% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scdblfinder.R
\name{add_doublets}
\alias{add_doublets}
\title{Add artificial doublets to Seurat object}
\usage{
add_doublets(
  obj,
  samples = obj@meta.data$orig.ident,
  clusters,
  prep_new_obj = T,
  ...
)
}
\arguments{
\item{obj}{seurat object}

\item{samples}{vector of cluster idents, e.g. orig.ident from obj;
obj is spit by that and doublets added separately; can be one sample only}

\item{clusters}{vector of cluster idents, e.g. a meta.col from obj}

\item{prep_new_obj}{make a new seurat to fit doublets into dim red}

\item{...}{args to SO_prep02}
}
\value{
seurat
}
\description{
Used within run_scdblfinder function.
Using scDblFinder::addDoublets with accessory lines of code.
}
