% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO_prep01.R
\name{add_pct_featset_and_cc}
\alias{add_pct_featset_and_cc}
\title{Add typical feature set pct and cell cycle score}
\usage{
add_pct_featset_and_cc(obj, species = "..auto..")
}
\arguments{
\item{obj}{seurat object}

\item{species}{human or mouse or ..auto.. for guessing}
}
\value{
obj with updated meta data
}
\description{
Add typical feature set pct and cell cycle score
}
