% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cluster_assignment_by_module_scores.R
\name{best_cluster_assignment_by_module_scores0}
\alias{best_cluster_assignment_by_module_scores0}
\title{Assign modules/gene sets/pathways to anonymous clusters in optimal way}
\usage{
best_cluster_assignment_by_module_scores0(df, score, module, cluster)
}
\arguments{
\item{df}{long data frame}

\item{score}{module score column}

\item{module}{module name column}

\item{cluster}{cluster column which module is to be mapped to}
}
\description{
Assign module scores that cover each cell with a reasonable maximum and
have re-calculated clustered assigned to them.
This function is called by discretize_module_score_suggestive_clusters().
}
