% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_dimred_to_seurat.R
\name{cds_dimred_to_seurat}
\alias{cds_dimred_to_seurat}
\title{Copy dimension reductions from a cell data set (cds) object to respective slots of a Seurat object (SO)}
\usage{
cds_dimred_to_seurat(
  cds,
  SO,
  reductions = NULL,
  suffix = "monocle",
  assay = "RNA"
)
}
\arguments{
\item{cds}{a cell data set}

\item{SO}{an Seurat object}

\item{reductions}{which reductions to copy; must be one or multiple of names(SingleCellExperiment::reducedDims(cds))}

\item{suffix}{which suffix to add to the reduction names in the Seurat object}

\item{assay}{which assay the dimension reduction were originally based on; must be one of names(SO@assays)}
}
\description{
cds may have been created by monocle3.
}
