% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_trajectory_to_seurat.R
\name{cds_trajectory_to_seurat}
\alias{cds_trajectory_to_seurat}
\title{Copy a trajectory from a cell data set (cds) object to respective the Misc slot of a Seurat object (SO)}
\usage{
cds_trajectory_to_seurat(
  cds,
  SO,
  x = 1,
  y = 2,
  reduction_method = "UMAP",
  name = "trajectory"
)
}
\arguments{
\item{cds}{a cell data set}

\item{SO}{an Seurat object}

\item{name}{}
}
\description{
The code lines are copied from monocle3::plot_cells
The aim is to have all data in a Seurat object to allow plot creation with a common style/layout.
}
