% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_stats.R
\name{coexpression_stats}
\alias{coexpression_stats}
\title{Inspect reference feature expression in groups of feature coexpressers}
\usage{
coexpression_stats(obj, features, ref_feature, metric = "coexpr_sum")
}
\arguments{
\item{obj}{seurat object}

\item{features}{features, the groups of which are assessed for ref_feature}

\item{ref_feature}{reference feature}

\item{metric}{metric returned from scexpr::coexpression_metrics}
}
\description{
Inspect reference feature expression in groups of feature coexpressers
}
\examples{
\dontrun{
coexpression_stats(so, c("CD53", "CD70", "CD27"), "KLF2")
}
}
