% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_ident.R
\name{convert_gene_ident}
\alias{convert_gene_ident}
\title{Convert gene identifiers}
\usage{
convert_gene_ident(
  x,
  input = c("hgnc_symbol", "external_synonym", "external_gene_name"),
  output = c("ensembl_gene_id", "hgnc_symbol", "external_gene_name", "external_synonym",
    "gene_biotype", "description")
)
}
\arguments{
\item{x}{input identifiers}

\item{input}{input ident, filters in biomaRt}

\item{output}{output ident, attributes in biomaRt}
}
\value{
list()
}
\description{
Uses biomaRt. See biomaRt::listFilters() and biomaRt::listAttributes()
}
\examples{
convert_gene_ident(x = c("CD8A", "CD4"))
}
