% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowplot2.R
\name{elbowplot2}
\alias{elbowplot2}
\title{Plot variance along PCs}
\usage{
elbowplot2(
  obj,
  assay = "RNA",
  npcs = min(30, length(Seurat::VariableFeatures(obj[[1]]))),
  theme = ggplot2::theme_grey()
)
}
\arguments{
\item{obj}{single seurat or list thereof}

\item{assay}{assay}

\item{npcs}{how many PC to print}

\item{theme}{ggplot theme}
}
\value{
list of data and plots
}
\description{
Plot variance along PCs
}
