% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{feature_plot}
\alias{feature_plot}
\title{Title}
\usage{
feature_plot(
  SO,
  features,
  assay = c("RNA", "SCT"),
  dims = c(1, 2),
  cells = NULL,
  downsample = 1,
  make.cellnames.unique = F,
  pt.size = 1,
  pt.size.expr.factor = 1,
  order = T,
  min.q.cutoff = 0,
  max.q.cutoff = 1,
  reduction = "tSNE",
  split.by = NULL,
  shape.by = NULL,
  combine = T,
  ncol.combine = NULL,
  nrow.combine = NULL,
  feature.aliases = NULL,
  title = NULL,
  cutoff.feature = NULL,
  cutoff.expression = 0,
  exclusion.feature = NULL,
  binary.expr = F,
  col.expresser = "#E69F00",
  legend.position = "right",
  legend.title.text.size = 14,
  legend.barheight = 3,
  legend.barwidth = 0.5,
  legend.text.size = 10,
  legend.nrow = NULL,
  legend.ncol = NULL,
  legend.key.size = 0.3,
  shape.legend.size = 3,
  hide.shape.legend = F,
  col.legend.size = 3,
  col.pal.c = "spectral",
  col.pal.d = "custom",
  col.excluded.cells = "#ededed",
  col.non.expresser = "#cfcfcf",
  col.pal.rev = F,
  theme = NULL,
  plot.axis.labels = F,
  plot.panel.grid = F,
  title.font.size = 14,
  plot.freq.of.expr.title = T,
  plot.freq.of.expr.annotation = T,
  plot.legend.title = F,
  plot.title = T,
  plot.order = NULL,
  annotation.position = c(-Inf, Inf),
  annotation.font.size = 4,
  plot.cutoff.feature.inset = F,
  inset.position = c(0.75, 0.1),
  inset.size = c(0.15, 0.15),
  strip.font.size = 14,
  plot.strip.selection = NA,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
