% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_pairwise.R
\name{find_markers_pairwise}
\alias{find_markers_pairwise}
\title{Find marker genes in pairwise comparisons}
\usage{
find_markers_pairwise(
  obj,
  group = NULL,
  split = NULL,
  mc.cores = 1,
  method = c("scexpr", "seurat"),
  redundant = F,
  ...
)
}
\arguments{
\item{obj}{Seurat object}

\item{group}{column in meta data}

\item{split}{column in obj to split by before marker calculation}

\item{mc.cores}{multiple threads?}

\item{method}{use scexpr::find_all_marker or Seurat::FindMarkers}

\item{redundant}{return redundant results?}

\item{...}{arguments to Seurat::FindMarkers}
}
\value{
data.frame
}
\description{
Find marker genes in pairwise comparisons
}
\examples{
\dontrun{
markers_pairwise <- scexpr::find_markers_pairwise(so, group = "label_main", mc.cores = 6)

# turn this into a result like from SingleR::getClassicMarkers but based on more statistic
markers_pairwise_seurat <- markers_pairwise |>
  dplyr::slice_min(order_by = p_val, n = 100, by = c(ident.1, ident.2), with_ties = F) |>
  dplyr::arrange(p_val)
markers_pairwise_seurat <- split(markers_pairwise_seurat, markers_pairwise_seurat$ident.1)
markers_pairwise_seurat <- purrr::map(markers_pairwise_seurat, ~split(.x$gene, .x$ident.2))
markers_pairwise_seurat_full <- scexpr:::make_list_of_lists_redundant(markers_pairwise_seurat)
}
}
