% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_cycle_genesets.R
\name{get_cell_cycle_genesets}
\alias{get_cell_cycle_genesets}
\title{Get cell cycle markers from Seurat and Whitfield and scran}
\usage{
get_cell_cycle_genesets()
}
\value{
list of different cell cycle associated gene sets
}
\description{
Get cell cycle markers from Seurat and Whitfield and scran
}
\examples{
ccgenelsz <- get_cell_cycle_genesets()
}
