% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_GO_terms.R
\name{get_GO_meta}
\alias{get_GO_meta}
\title{Meta data from GO sets}
\usage{
get_GO_meta(goid)
}
\arguments{
\item{goid}{GO id(s)}
}
\value{
list, one entry for each goid
}
\description{
Meta data from GO sets
}
\examples{
\dontrun{
get_GO_meta(c("GO:0044281", "GO:0006082", "GO:0006520", "GO:0032787", "GO:0044283",
              "GO:0006629", "GO:0016054", "GO:0044282", "GO:0006631", "GO:0016053"))
}
}
