% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_GO_terms.R
\name{get_GO_parents}
\alias{get_GO_parents}
\title{Get parents of GO.db id}
\usage{
get_GO_parents(goid)
}
\arguments{
\item{goid}{GO id(s)}
}
\value{
data frame
}
\description{
Get parents of GO.db id
}
\examples{
\dontrun{
get_GO_parents(c("GO:0006084", "GO:0015800"))
c5 <- msigdbr::msigdbr(collection = "C5", subcollection = "BP")
go <- unique(c5$gs_exact_source)
get_GO_parents(go[1])
#' }
}
