% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_GO_terms.R
\name{get_GO_parents_recursive}
\alias{get_GO_parents_recursive}
\title{Get parents of GO.db id recursively}
\usage{
get_GO_parents_recursive(goid, seen = character())
}
\arguments{
\item{goid}{GO id vector}

\item{seen}{used by recursion, leave empty}
}
\value{
from-to-dataframe
}
\description{
Build a graph from it.
}
\examples{
\dontrun{
c5 <- gsea_get_msigdb("C5")
c5 <- msigdbr::msigdbr(collection = "C5", subcollection = "BP")
go <- unique(c5$gs_exact_source)
goterm_fromtodf <- get_GO_parents_recursive(go[1:3])
plot_GO_graph(fromtodf = goterm_fromtodf)
}
}
