% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_groupwise.R
\name{gsea_groupwise}
\alias{gsea_groupwise}
\title{Groupwise gene set enrichment analysis}
\usage{
gsea_groupwise(
  obj,
  group,
  fgseaMultilevel_args = list(pathways = list()),
  get_layer_args = list()
)
}
\arguments{
\item{obj}{Seurat object}

\item{group}{grouping column in obj meta.data}

\item{fgseaMultilevel_args}{arguments to fgsea::fgseaMultilevel; named list of
pathways is mandatory}

\item{get_layer_args}{arguments to scexpr::get_layer}
}
\value{
list
}
\description{
S2N is calculated with scexpr::gsea_s2n_groupwise, then fgsea::fgseaMultilevel is
called on each group. You may use an overclustered grouping, GSEA on groups of cells
should cancel out noise in single cells and group highly correlates cells.
}
