% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_inspect_ranked_list.R
\name{gsea_inspect_ranked_list}
\alias{gsea_inspect_ranked_list}
\title{Check and visualize ties in ranked gene list}
\usage{
gsea_inspect_ranked_list(x)
}
\arguments{
\item{x}{numeric vector to check for ties}
}
\value{
list with table, data.frame and ggplots
}
\description{
Check and visualize ties in ranked gene list
}
