% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plot_summary.R
\name{gsea_plot_summary}
\alias{gsea_plot_summary}
\title{Plot summary statistics of many enriched gene sets, i.e. enrichplot}
\usage{
gsea_plot_summary(
  gsea_df,
  x = c("NES", "padj", "leadingEdge_size", "leadingEdge_size_rel", "ES"),
  y = "pathway",
  fill = c("padj", "NES", "leadingEdge_size", "leadingEdge_size_rel", "ES"),
  size = c("leadingEdge_size_rel", "leadingEdge_size", "padj", "NES", "ES"),
  geom = c("point", "col"),
  add_size_to_y = T,
  minus_log10_padj = T,
  order_y_by = x,
  size_range = c(3, 10),
  x_expand_mult = c(0.1, 0.1),
  theme = colrr::theme_material(white = T),
  col_pal = colrr::col_pal(name = "spectral", direction = -1),
  title = NULL
)
}
\arguments{
\item{gsea_df}{(subset) data frame returned from gsea_on_msigdbr}

\item{x}{variable/column on x axis}

\item{y}{variable/column on y axis}

\item{fill}{variable/column for fill}

\item{size}{variable/column for dot size}

\item{geom}{which geom to plot}

\item{add_size_to_y}{add size of gene set to y axis text}

\item{minus_log10_padj}{calculate -log10 of adjusted p-value}

\item{order_y_by}{which column to order y axis breaks by}

\item{size_range}{dot size range from to}

\item{x_expand_mult}{x axis expansion (mult) to accommodate dots fully}

\item{theme}{ggplot theme}

\item{col_pal}{color palette for continuous fill}

\item{title}{plot title}
}
\value{
ggplot
}
\description{
Plot summary statistics of many enriched gene sets, i.e. enrichplot
}
\examples{
\dontrun{
gsea <- scexpr::gsea_on_msigdbr(gene_ranks = gene_ranks,
                                use_msigdbr = T,
                                fgsea_args = list(minSize = 10),
                                msigdbr_args = list(db_species = "HS",
                                                    species = "human",
                                                    collection = "C5"))
gsea_df <- gsea[["data"]]
gsea_plot_summary(gsea_df = gsea_df |>
                    dplyr::filter(NES>0) |>
                    dplyr::slice_min(padj, n = 10))
}
}
