% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_plot_summary.R
\name{gsea_plot_summary_default_groups}
\alias{gsea_plot_summary_default_groups}
\title{Call gsea_plot_summary with pre-defined groups}
\usage{
gsea_plot_summary_default_groups(
  gsea_df,
  topn = 10,
  metric = c("padj", "NES", "leadingEdge_size", "leadingEdge_size_rel", "ES"),
  join_NES_plus_minus = F,
  split = c("gs_collection", "gs_subcollection"),
  sep = "_",
  ...
)
}
\arguments{
\item{gsea_df}{data frame returned from gsea_on_msigdbr}

\item{topn}{how many gene sets per group, no ties}

\item{metric}{what to sort by for topn}

\item{join_NES_plus_minus}{enriched and de-enriched sets are plotted
separately by default; join instead}

\item{split}{which columns to split by}

\item{sep}{separator for split labels}

\item{...}{args to gsea_plot_summary}
}
\value{
list of filtered split data frames and ggplots
}
\description{
Goal is to get a structured overview.
}
\examples{
\dontrun{
gsea <- gsea_on_msigdbr(gene_ranks = gene_ranks,
                        use_msigdbr = T,
                        fgsea_args = list(minSize = 10),
                        msigdbr_args = list(db_species = "HS",
                                            species = "human",
                                            collection = "C5"))
gsea_df <- gsea[["data"]]
grouped_plots <- gsea_plot_summary_default_groups(gsea_df)
grouped_plots[["plots"]][["C5_GO:BP___plus"]]

# plot graph of GO gene sets
goid <- grouped_plots[["dfs"]][["C5_GO:BP___plus"]]$gs_exact_source
plot_GO_graph(goid = goid)
plot_GO_graph(goid = goid,
              layout = "dendrogram",
              multiline = "long")
}
}
