% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_material.R
\name{guides_default}
\alias{guides_default}
\title{Add default legend guide elements to ggplot}
\usage{
guides_default(colorbar = F, ...)
}
\arguments{
\item{colorbar}{plot color legend as bar}

\item{...}{arguments to guide_colorbar or guide_legend}
}
\description{
Make default barheight = 10 and width = 1, legend title as
ggtext::element_markdown, point legends of size = 4. Valid for fill and color.
}
