% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pseudobulk.R
\name{heatmap_pseudobulk}
\alias{heatmap_pseudobulk}
\title{Title}
\usage{
heatmap_pseudobulk(
  SO,
  assay = c("RNA", "SCT"),
  meta.col = NULL,
  levels.calc = NULL,
  levels.plot = NULL,
  features = NULL,
  normalization = c(-1, 1),
  topn.features = 10,
  topn.metric = c("logFC", "auc", "padj"),
  min.pct = 0.1,
  max.padj = 0.05,
  title = NULL,
  title.font.size = 14,
  font.family = "sans",
  y.font.size = 10,
  tile.borders = T,
  dotplot = F,
  plot.feature.breaks = T,
  plot.sec.axis = F,
  legend.position = "right",
  legend.direction = "vertical",
  legend.barheight = 8,
  legend.barwidth = 1,
  legend.text.size = 10,
  legend.title.text.size = 10,
  legend.title.fill = NULL,
  legend.title.size = NULL,
  legend.title.position = "top",
  legend.labels = c("min", "", "max"),
  feature.labels = NULL,
  feature.labels.nudge_x = -0.1,
  feature.labels.axis.width = 0.2,
  ...
)
}
\arguments{
\item{levels.calc}{which levels in meta.col to include in calculation; all levels if NULL; also defines order on x-axis}

\item{levels.plot}{which levels in meta.col (and subset of levels.calc) to include in platting; set to levels.calc if NULL; also defines order on x-axis}

\item{...}{}
}
\value{

}
\description{
Title
}
