% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_typing.R
\name{hla_typing}
\alias{hla_typing}
\title{Title}
\usage{
hla_typing(
  hla_ref,
  reads,
  diff_allele_hits_single_results = 5,
  lapply_fun = lapply
)
}
\arguments{
\item{lapply_fun}{}
}
\value{

}
\description{
Title
}
