% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_mm_ortholog_table.R
\name{hs_mm_ortholog_table}
\alias{hs_mm_ortholog_table}
\title{Create an table of ortholog genes between human and mouse}
\usage{
hs_mm_ortholog_table(features_hs, features_mm)
}
\arguments{
\item{features_hs}{vector of human genes (gene symbols, hgnc)}

\item{features_mm}{vector of human genes (gene symbols)}
}
\value{
a list with (i) a table of orthologs and (ii, iii) vectors of genes which no orthologs were found for
}
\description{
A combined table with entries from ProjecTILs::Hs2Mm.convert.table and
entries from a bioMart query is generated. features_hs is used
to query bioMart. So, it is intended to look for mouse orthologs based
on human features.
}
