% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineplot_avg_expr.R
\name{lineplot_avg_expr}
\alias{lineplot_avg_expr}
\title{Make a lineplot of average gene expression across groups}
\usage{
lineplot_avg_expr(
  obj,
  features,
  assay = "RNA",
  group = "orig.ident",
  split = NULL,
  return_df = F
)
}
\arguments{
\item{obj}{seurat object}

\item{features}{gene features}

\item{assay}{which assay}

\item{group}{groups on x-axis}

\item{split}{another group as color scale}

\item{return_df}{}
}
\value{
ggplot
}
\description{
Make a lineplot of average gene expression across groups
}
\examples{
\dontrun{
gg <- lineplot_avg_expr(obj = so, features = c("CD3", "NKG7"), split = "cluster")
}
}
