% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb_get.R
\name{msigdb_get}
\alias{msigdb_get}
\title{Convenience function to get gene sets from msigdbr}
\usage{
msigdb_get(
  collection = c("C1", "C2", "C3", "C4", "C5", "C6", "C7", "C8", "H"),
  msigdbr_args = list(db_species = "HS", species = "human"),
  return = c("list", "df"),
  return_df = c("reduced", "full")
)
}
\arguments{
\item{collection}{vector of collections}

\item{msigdbr_args}{arguments to msigdbr::msigdbr}

\item{return}{what to return; df for manual inspect, list for passing to gsea
functions}

\item{return_df}{when return df, reduce columns to gs_collection, gs_name,
gene_symbol?}
}
\value{
list of gene sets and data frame with category info
}
\description{
Convenience function to get gene sets from msigdbr
}
\examples{
sets <- msigdb_get(collection = c("C1", "H"))
# then pass on
fgsea_groupwise(obj = SO,
                group = "meta.col",
                fgseaMultilevel_args = list(pathways = list(sets$sets)))
}
