% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_GO_terms.R
\name{plot_GO_graph}
\alias{plot_GO_graph}
\title{Plot graph/network of GO terms}
\usage{
plot_GO_graph(
  goid = "",
  fromtodf = NULL,
  layout = "tree",
  label_plot = T,
  label_repel = F,
  label_multiline = T,
  multiline = c("wide", "long"),
  label_goid = F,
  return_graph = F,
  expand_x = c(0.3, 0.3),
  expand_y = c(0.1, 0.1),
  label_split = " |_|-",
  ...
)
}
\arguments{
\item{goid}{GO id(s)}

\item{fromtodf}{get_GO_parents_recursive(goid) instead of goid; modify optionally}

\item{layout}{layout of ggraph,
e.g. auto, tree, dendrogram, circlepack, htree, cactustree, treemap, partition,
fr, kk, drl, graphopt, lgl, stress, circle, linear, focus, centrality,
grid, star, sphere, random, hive, sugiyama}

\item{label_plot}{plot labels?}

\item{label_repel}{repel labels?}

\item{label_multiline}{wrap labels to multiple lines}

\item{multiline}{split at every space (long) or at second space only (wide)}

\item{label_goid}{label GO id instead of term?}

\item{return_graph}{return graph instead of ggplot flr own plotting}

\item{expand_x}{mult x axis expansion}

\item{expand_y}{mult y axis expansion}

\item{label_split}{label split chars}

\item{...}{args to ggraph::geom_node_label}
}
\value{
ggplot
}
\description{
Filtered for isa-entries
}
\examples{
\dontrun{
c5 <- msigdbr::msigdbr(collection = "C5", subcollection = "BP")
go <- unique(c5$gs_exact_source)
plot_GO_graph(go, label_plot = F) # whole network
plot_GO_graph(go[1])
plot_GO_graph(go[1:2])
}
}
