% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Title}
\usage{
plot_gsea(
  data,
  ticksSize = 0.2,
  color_ES_line = "black",
  color_min_max_ES_line = "black",
  color_leadingEdge = "hotpink2",
  label_genes = NULL,
  theme = ggplot2::theme_bw(),
  plot_leadingEdge_rank = T,
  plot_leadingEdge_size = T
)
}
\arguments{
\item{plot_leadingEdge_size}{}
}
\description{
Title
}
