% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data_df_wide.R
\name{prep_data_df_wide}
\alias{prep_data_df_wide}
\title{Get data frame for plotting from Seurat object(s)}
\usage{
prep_data_df_wide(
  data,
  feature,
  reduction = c("umap_1", "umap_2"),
  qmin = 0,
  qmax = 1,
  order = T,
  order_rev = F,
  order_abs = T,
  order_discr = T,
  shuffle = F,
  bury_NA = T,
  na_rm = F,
  inf_rm = F,
  label_feature = NULL,
  contour_feature = NULL,
  split_feature = NULL,
  shape_feature = NULL
)
}
\arguments{
\item{feature}{columns that are features to plot as color scale}

\item{reduction}{columns that are dimension reduction}

\item{qmin}{lower quantile of feature values where to cut the color scale}

\item{qmax}{upper quantile}

\item{order}{order rows from lowest to highest (negative at bottom)}

\item{order_rev}{reverse row order}

\item{order_abs}{absolute values for ordering: zeros first and anything
distant from zero on top}

\item{order_discr}{set TRUE to for order by abundance, only relevant for
discrete meta features}

\item{shuffle}{do shuffle rows (if order is FALSE) to randomize plotting
order; if !order and !shuffle the provided order is not altered}

\item{bury_NA}{put NA to the bottom which will bury them upon plotting}

\item{na_rm}{remove NA values? (only for meta feature)}

\item{inf_rm}{remove infinite values?}

\item{label_feature}{feature for plotting labels}

\item{contour_feature}{feature for plotting contours}

\item{split_feature}{feature for creating facets}

\item{shape_feature}{feature to shape plotted points by}
}
\value{
a list of data frames
}
\description{
From a wide data frame with any kind of expression values,
prepare a list of data frames ready to pass to feature_plot_data.
}
