% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SO.R
\name{prep_SO}
\alias{prep_SO}
\title{Prepare a somewhat standardized Seurat Object for further analyses}
\usage{
prep_SO(
  SO_unprocessed,
  samples,
  cells,
  min_cells = 50,
  downsample = 1,
  export_prefix = NULL,
  cluster_resolutions = 0.8,
  reductions = c("tsne", "umap"),
  nhvf = 800,
  npcs = 40,
  nintdims = 30,
  normalization = c("SCT", "LogNormalize"),
  batch_corr = c("harmony", "integration", "regression", "none"),
  ref_sample = NULL,
  integr_reduction = c("rpca", "cca", "rlsi"),
  vars.to.regress = NULL,
  seeed = 42,
  save_path = NULL,
  ...
)
}
\arguments{
\item{SO_unprocessed}{named list of split Seurat objects}

\item{samples}{names of SO_unprocessed that are to include; if missing all are used}

\item{cells}{vector of cell names to include; if missint all cells are used}

\item{min_cells}{min number of cells per object in SO_unprocessed; objects with lower cell number are excluded}

\item{downsample}{downsample a fraction of cells from each object in SO_unprocessed (downsample < 0);
or an absolute number of cells per object (downsample > 1 & downsample >= min_cells)}

\item{export_prefix}{prefix to the output rds file}

\item{cluster_resolutions}{vector of resolutions to compute by Seurat::FindClusters}

\item{reductions}{which reduction to calculate, tsne --> Seurat::RunTSNE, umap --> Seurat::RunUMAP}

\item{nhvf}{number high variables features, passed to Seurat::SCTransform
or Seurat::FindVariableFeatures}

\item{npcs}{number of principle components in calculate in pca and to
consider for downstream functions like tSNE or UMAP}

\item{nintdims}{number of integration dimension, when batch_corr = "integration"}

\item{normalization}{algorithm for normalization of UMIs}

\item{batch_corr}{algorithm for batch correction between samples in SO_unprocessed;
only relevant if more than 1 sample is passed}

\item{ref_sample}{optional names of the reference sample in SO_unprocessed for
Seurat::FindIntegrationAnchors or RunHarmony; reference; if NULL no reference is used}

\item{integr_reduction}{reduction method for Seurat::FindIntegrationAnchors; rpca is quicker than cca}

\item{vars.to.regress}{passed to Seurat::SCTransform or Seurat::ScaleData, applied if batch_corr == "regression"}

\item{seeed}{seed passed to several methods}

\item{save_path}{folder to save the resulting Seurat object as rds-file to}

\item{...}{additional arguments passed to harmony::RunHarmony, Seurat::RunUMAP,
Seurat::RunTSNE, Seurat::FindNeighbors, Seurat::FindClusters,
Seurat::PrepSCTIntegration, Seurat::FindIntegrationAnchors, Seurat::IntegrateData}
}
\value{

}
\description{
Prepare a somewhat standardized Seurat Object for further analyses
}
\examples{
\dontrun{
}
}
