% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads_from_bam.R
\name{reads_from_bam}
\alias{reads_from_bam}
\title{Get reads from a bam file}
\usage{
reads_from_bam(
  file_path,
  genomic_ranges,
  add_tags = c("CR", "CY", "AS", "UR", "UY", "HI", "NH", "nM", "RE"),
  add_flags = NA,
  read_scores = T,
  revcomp_minus_strand = T
)
}
\arguments{
\item{file_path}{path to a position-sorted bam file; index file (.bai) has to be in the same directory}

\item{genomic_ranges}{GRanges object with n genomic ranges}

\item{add_tags}{tags to extract from bam file, passed to Rsamtools::ScanBamParam()}

\item{add_flags}{a named list of columns to add; each element must have n elements (length of genomic_ranges)}

\item{read_scores}{calculate read scores from PhredQuality}

\item{revcomp_minus_strand}{calculate reverse complement of reads on minus strand}
}
\value{
a data frame with flagged (annotated) reads
}
\description{
to do
}
\details{
Read scores: https://support.illumina.com/help/BaseSpace_OLH_009008/Content/Source/Informatics/BS/QualityScoreEncoding_swBS.htm
CellRanger tags: Cell barcode (CR), Cell barcode read quality (CY), Alignment score (AS), UMI (UR),
UMI read quality (UR), Query hit index (HI), Number of reported alignments for query (NH), Number of mismatches per pair (nM),
Region type (E = exonic, N = intronic, I = intergenic) (RE)
}
