% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_reduction.R
\name{rename_reduction}
\alias{rename_reduction}
\title{Rename a reduction in Seurat at all necessary positions}
\usage{
rename_reduction(obj, reduction, new_name)
}
\arguments{
\item{obj}{seurat object}

\item{reduction}{reduction name (current)}

\item{new_name}{new name to assign}
}
\value{
seurat object
}
\description{
Rename a reduction in Seurat at all necessary positions
}
\examples{
\dontrun{
so <- rename_reduction(obj = so, reduction = "XUMAP" new_name = "umap")
}
}
