% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monocle.R
\name{run_monocle}
\alias{run_monocle}
\title{Calculate monocle trajectory and add to seurat object}
\usage{
run_monocle(obj)
}
\arguments{
\item{obj}{seurat object with umap dr}
}
\value{
seurat object
}
\description{
Calculate monocle trajectory and add to seurat object
}
\examples{
\dontrun{
scexpr::feature_plot2(obj, "umap_partition", "umap") +
  ggnewscale::new_scale_color() +
  ggplot2::geom_segment(ggplot2::aes(x=source_prin_graph_dim_1,
                                     y=source_prin_graph_dim_2,
                                     xend=target_prin_graph_dim_1,
                                     yend=target_prin_graph_dim_2,
                                     color = source_umap_partition),
                        size=0.75,
                        linetype="solid",
                        na.rm=TRUE,
                        data=obj@misc[["umap_edge_df"]])
}
}
