% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scdblfinder.R
\name{run_scdblfinder}
\alias{run_scdblfinder}
\title{Run scDblFinder with accessory code lines}
\usage{
run_scdblfinder(
  obj,
  assay = "RNA",
  features = scexpr::get_var_features(so),
  samples_col = "orig.ident",
  clusters_col = "seurat_clusters",
  min_umi = 300,
  min_cells_per_group = 20,
  dims = ncol(obj@reductions$pca@cell.embeddings),
  add_doublets = F,
  add_doublets_tell = F
)
}
\arguments{
\item{obj}{seurat object}

\item{assay}{which assay}

\item{features}{gene features to use}

\item{samples_col}{samples to treat separately, vector of idents}

\item{clusters_col}{cluster dents to help the algorithm and optionally to
add_doublets}

\item{min_umi}{minimum total UMI sum across features per cell; cells below
are not considered for doublet inference}

\item{dims}{arg to scDblFinder::scDblFinder}

\item{add_doublets}{add artificial doublets}

\item{add_doublets_tell}{tell scDblFinder::scDblFinder about artificial doublets}
}
\value{
seurat
}
\description{
Run scDblFinder with accessory code lines
}
\examples{
\dontrun{
so <- run_scdblfinder(so)
}
}
