% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_soupx2.R
\name{run_soupx2}
\alias{run_soupx2}
\title{Run default SoupX pipeline}
\usage{
run_soupx2(
  filt,
  raw,
  obj = NULL,
  so_prep_args = list(FindClusters_args = list(resolution = 0.8), nhvf = 2000, npcs = 20),
  autoEstCont_args = list(),
  adjustCounts_method = "subtraction"
)
}
\arguments{
\item{filt}{filtered gene x cell matrix or list thereof; when list: everything
is treated as one}

\item{raw}{raw gene x cell matrix or list thereof}

\item{obj}{provide seurat object instead of filt}

\item{so_prep_args}{args to SO_prep03/SO_prep02}

\item{autoEstCont_args}{args to SoupX::autoEstCont}

\item{adjustCounts_method}{to method in SoupX::adjustCounts; one of
c("subtraction", "soupOnly", "multinomial")}
}
\value{
Seurat object with SOUPX corrected counts
}
\description{
Run default SoupX pipeline
}
\examples{
\dontrun{
filt <- Seurat::Read10X_h5("filtered_feature_bc_matrix.h5")
raw <- Seurat::Read10X_h5("raw_feature_bc_matrix.h5")
so <- run_soupx2(filt, raw)
}
}
