% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhoutte_plot.R
\name{silhoutte_plot}
\alias{silhoutte_plot}
\title{Make silhouette plot for clustering}
\usage{
silhoutte_plot(
  obj,
  clustering,
  reduction = "pca",
  theme = colrr::theme_material(white = T),
  col_pal = colrr::col_pal("custom")
)
}
\arguments{
\item{obj}{seurat object}

\item{clustering}{column in meta.data}

\item{reduction}{reduction name}

\item{theme}{ggplot theme}

\item{col_pal}{color and fill palette}
}
\value{
ggplot and data frame to misc slot
}
\description{
From Seurat object. Data frame added to misc slot.
}
