% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO_prep03.R
\name{SO_prep03}
\alias{SO_prep03}
\title{Prepare Seurat from gene x cell matrix with quick defaults}
\usage{
SO_prep03(
  matrix_list,
  normalization = "LogNormalize",
  nhvf = 800,
  npcs = 20,
  interactive_varfeat_selection = F,
  interactive_pc_selection = F,
  ...
)
}
\arguments{
\item{matrix_list}{list of gene x cell matrices, e.g. from read_10X_data}

\item{normalization}{which normalization; LogNormalize is quicker}

\item{nhvf}{n var features}

\item{npcs}{n principle components}

\item{interactive_varfeat_selection}{see SO_prep02}

\item{interactive_pc_selection}{see SO_prep02}

\item{...}{args to SO_prep02}
}
\value{
seurat object
}
\description{
SO_prep02 is called.
}
