% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_material.R
\name{theme_material}
\alias{theme_material}
\title{ggplot2 theme with style of Rstudios color palette 'Material'}
\usage{
theme_material(
  base_size = 14,
  text_color = "white",
  bg_color = "#273238",
  bg_color2 = "#4F5C69",
  text_fun = ggtext::element_markdown,
  style = c("prism", "prismy", "prismx", "base"),
  white = F,
  legend_tight = F
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{text_color}{color of various text elements}

\item{bg_color}{background color, not intended to be changed dramatically}

\item{bg_color2}{background color 2, e.g. for facet strip}

\item{text_fun}{function for text elements, ggplot2::element_text or
ggtext::element_markdown}

\item{style}{appearance of axes and grid}

\item{white}{use a light color scale with white bg}

\item{legend_tight}{legend more compact and closer to plot panel}

\item{...}{}
}
\description{
ggplot2 theme with style of Rstudios color palette 'Material'
}
