% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfeat_plot.R
\name{varfeat_plot}
\alias{varfeat_plot}
\title{Plot overview of different nHVF}
\usage{
varfeat_plot(obj, n_varfeat = seq(200, 2000, 200))
}
\arguments{
\item{obj}{seurat object}

\item{n_varfeat}{number of variable feature to visualize (cutoffs)}
}
\value{
patchwork of ggplots
}
\description{
Plots mean-variance plots with different cutoffs. Variable features need
to be calculated in obj before calling this function.
}
\examples{
seu <- readRDS(system.file("extdata", "SO_5k_pbmc_v3_RNA_none_1_800_12_small.rds", package = "scexpr"))
out <- varfeat_plot(seu) # a bit stupid because all are var_feat
}
