% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano01_calc.R
\name{volcano01_calc}
\alias{volcano01_calc}
\title{Volcano plot comparing gene expression between groups of cells from one or more Seurat objects}
\usage{
volcano01_calc(
  SO,
  assay = "RNA",
  layer = "data",
  neg_cells,
  pos_cells,
  neg_name = NULL,
  pos_name = NULL,
  add_n_repl_to_names = F,
  method = c("MAST", "wilcox", "wilcox_limma", "bimod", "roc", "t", "negbinom",
    "poisson", "LR", "DESeq2", "limma", "custom"),
  fc_inf_squish = 2,
  p_zero_squish = 1,
  min_pct = 0.1,
  p_adj = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr", "none"),
  fc_thresh = 0,
  feature_plot_args = list(col_pal_d_args = list(name =
    stats::setNames(colrr::col_pal("custom")[2:3], c(neg_name, pos_name)),
    missing_fct_to_na = F)),
  mc.cores = 1,
  species = "Hs",
  ...
)
}
\arguments{
\item{SO}{one or more Seurat object(s); provide multiple objects as a (named) list}

\item{assay}{which assay to get expression data from}

\item{layer}{which data layer}

\item{neg_cells}{vector of cell names for the negative group of cells; genes which
are expressed at higher level in this group will have a negative sign (minus); use Seurat::Cells()
or Seurat::WhichCells() to select or use filter operations on SO@meta.data to select cells;
or provide a vector which will be checked against Idents(SO)}

\item{pos_cells}{vector of cell names for the positive group of cells}

\item{neg_name}{name for the negative group of cells; this will appear on plot axes}

\item{pos_name}{name for the positive group of cells; this will appear on plot axes}

\item{add_n_repl_to_names}{add number of (pseudo-) replicates to names}

\item{method}{backend for DE calculation; any method from Seurat::FindMarkers;
limma: intended for subsequent use MetaVolcanoR
custom: matrixTests::row_wilcoxon_twosample w/o any modelling}

\item{fc_inf_squish}{numeric; genes with an infinite fold-change (happens when a gene not expressed at all in one group)
are not left infinite but will be put at: "largest finite fold-change +/- fc_inf_squish"; such genes are indicated
specifically}

\item{p_zero_squish}{scales::oob_squish(df$p_val_adj, range = c(min(df$p_val_adj)/10^p_zero_squish, 1))}

\item{min_pct}{numeric; only consider genes for the volcano plot which are expressed at least by this
fraction of cells at least one group (neg_cells or pos_cells)}

\item{p_adj}{method for p-value adjustment}

\item{fc_thresh}{logfc.threshold in Seurat::FindMarkers}

\item{feature_plot_args}{arguments to feature_plot2}

\item{mc.cores}{mc cores for parallel computing when method is MAST}

\item{...}{arguments to Seurat::FindMarkers}
}
\value{
list of data frame, matrix, meta data; pct column in df is percent of
expressers for a gene in pos_group when avg_log2fc > 0 and vice versa pct
in neg_group when  avg_log2fc < 0
}
\description{
Calculate with volcano01_calc. Plot with volcano02_plot.
}
\examples{
\dontrun{
# when neg_cells and pos_cells are idents and not vectors of cell names
# set the right meta.data column as Idents first
SeuratObject::Idents(so2) <- so2$SCT_snn_res.0.8
  volc <- scexpr::volcano01_calc(
    so2,
    neg_cells = c(15,16),
    pos_cells = c(17,18),
    method = "wilcox")
  volcano02_plot(volc[["df"]])
}
}
