% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano02_plot.R
\name{volcano02_plot}
\alias{volcano02_plot}
\title{Title}
\usage{
volcano02_plot(
  vd,
  x = "log2.fc",
  y = "adj.p.val",
  feature = "feature",
  pt_size = 1,
  pt_alpha = 0.8,
  minus_log10_y = T,
  x_symm = T,
  x_expnd = 0.5,
  y_pseudo_log = F,
  pseudo_log_sigma = 1,
  features.to.color = NULL,
  features.color.by = NULL,
  errorbar.low.col = NULL,
  errorbar.up.col = NULL,
  errorbar.size = 0.2,
  errorbar.width = 0.2,
  col.pal = "RdBu",
  col.pal.dir = 1,
  col.type = c("c", "d"),
  font.size = 14,
  p_tick = NULL,
  feature_exclude = NULL,
  min_pct = 0,
  cut_p = NA,
  cut_fc = NA,
  label_plot = T,
  label_features = NULL,
  label_topn_metric = c("p_val", "fc", "both"),
  label_topn = 30,
  label_dot_color = "tomato2",
  label_neg_pos_sep = F,
  label_p_signif = 0.001,
  geom_text_repel_args = list(fontface = "italic", max.overlaps = 50)
)
}
\arguments{
\item{cut_fc}{}
}
\description{
Title
}
