% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano03_avg.R
\name{volcano03_avg}
\alias{volcano03_avg}
\title{Summarize DEG from multiple comparisons for volcano plot}
\usage{
volcano03_avg(volc01_df_list, neg_name, pos_name, logfc_avg_fun = mean)
}
\arguments{
\item{volc01_df_list}{list of data frames from scexpr::volcano01_calc}

\item{neg_name}{Name of the negative group. If \code{NULL}, the name is
inferred from the first data frame in \code{volc01_df_list}.}

\item{pos_name}{see neg_name}

\item{logfc_avg_fun}{function for averaging log fold changes}
}
\value{
data frame
}
\description{
Makes a data frame to use with volcano02_plot.
}
