% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_color_to_misc.R
\name{add_group_color_to_misc}
\alias{add_group_color_to_misc}
\title{Title}
\usage{
add_group_color_to_misc(
  obj,
  group = "orig.ident",
  colors = colrr::col_pal("custom"),
  name = paste0(group, "_colors")
)
}
\arguments{
\item{name}{}
}
\description{
Title
}
