% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds_trajectory_to_seurat.R
\name{cds_trajectory_to_seurat}
\alias{cds_trajectory_to_seurat}
\title{Copy a trajectory from a cell data set (cds) object to respective the Misc slot of a Seurat object (SO)}
\usage{
cds_trajectory_to_seurat(
  cds,
  SO,
  x = 1,
  y = 2,
  reduction_method = "UMAP",
  name = "trajectory",
  pseudotime_metacol_name = "pseudotime"
)
}
\arguments{
\item{cds}{a cell data set}

\item{SO}{an Seurat object}

\item{x}{column number of reduction_method to use as x-axis}

\item{y}{column number of reduction_method to use as y-axis}

\item{reduction_method}{name of reduction, currently only UMAP allows, according to monocle3}

\item{name}{target slot name in Seurat::Misc; will be overwritten if already existing}

\item{pseudotime_metacol_name}{name of new column in SO@meta.data for pseudotime information; set to NULL to not add a column to meta.data of SO}
}
\description{
The code lines are copied from monocle3::plot_cells
The aim is to have all data in a Seurat object to allow plot creation with a common style/layout.
}
