% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation2.R
\name{cluster_correlation2}
\alias{cluster_correlation2}
\title{Check relationship (correlation) of clusters from one or two Seurat objects}
\usage{
cluster_correlation2(
  objs,
  meta_cols,
  split = NULL,
  features = c("all", "pca"),
  assay = "RNA",
  layer = "data",
  method = c("pearson", "spearman", "kendall", "kendall_zi"),
  heatmap_long_df_args = list(fill = colrr::col_pal("spectral", direction = -1),
    colorsteps = "..auto.."),
  heatmap_ordering_args = list(feature_order = "hclust", group_order = "hclust"),
  lower_tri = F,
  min_cells = 1,
  avg_expression_args = list(fun = Matrix::rowMeans, fun2 = base::identity)
)
}
\arguments{
\item{objs}{one Seurat or a named list of 2 Seurats}

\item{meta_cols}{one or two columns from meta.data}

\item{split}{split correlation estimation by another variable like
donor: column from meta data; must exist in both objs with shared levels
when objs is a list of two}

\item{features}{which features (genes) to use for correlation: all intersecting,
intersecting pca features, or a custom vector}

\item{assay}{assay to pull expression matrix from}

\item{layer}{layer to pull expression matrix from}

\item{method}{correlation metric: from psych pkg: pearson, spearman, kendall;
from dismay}

\item{heatmap_long_df_args}{arguments to fcexpr::heatmap_long_df for plotting
results}

\item{heatmap_ordering_args}{arguments to fcexpr::heatmap_ordering}

\item{lower_tri}{for a symmetric matrix when only one obj is provided:
plot the lower triangle only? (remove redundancy)}

\item{min_cells}{min cell number in both groups to calc correlation}

\item{avg_expression_args}{args to scexpr::avg_expression}
}
\description{
Check https://github.com/skinnider/dismay and https://github.com/tpq/propr
propr could not be made for 2 matrices yet.
}
