% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_correlation_matrix.R
\name{cluster_correlation_matrix}
\alias{cluster_correlation_matrix}
\title{Title}
\usage{
cluster_correlation_matrix(
  SO,
  assay = c("RNA", "SCT"),
  meta.cols,
  levels = NULL,
  complement.levels = F,
  features = NULL,
  avg.expr,
  method = c("pearson", "kendall", "spearman"),
  corr.in.percent = FALSE,
  round.corr = 2
)
}
\arguments{
\item{round.corr}{}
}
\value{

}
\description{
Title
}
