% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gene_sets_by_leadingEdge.R
\name{cluster_gene_sets_by_leadingEdge}
\alias{cluster_gene_sets_by_leadingEdge}
\title{Title}
\usage{
cluster_gene_sets_by_leadingEdge(
  gsea_data_df,
  umap_args = list(),
  FindNeighbors_args = list(),
  FindClusters_args = list(resolution = 0.6),
  seed = 42
)
}
\arguments{
\item{FindClusters_args}{}
}
\description{
Title
}
