% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_volcanos.R
\name{clustering_volcanos}
\alias{clustering_volcanos}
\title{Title}
\usage{
clustering_volcanos(
  SO,
  assay = c("RNA", "SCT"),
  cluster.meta.cols,
  reduction = "tSNE",
  clustree.prefix = NULL,
  save.path = NULL,
  lapply_fun = lapply,
  ...
)
}
\arguments{
\item{assay}{}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; e.g. mc.cores may be passed when parallel::mclapply is chosen above}
}
\value{

}
\description{
Title
}
