% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_metrics.R
\name{coexpression_metrics}
\alias{coexpression_metrics}
\title{Measures of coexpression for a set of features}
\usage{
coexpression_metrics(
  obj,
  features,
  assay = "RNA",
  layer = "data",
  nameprefix = "coexpr",
  group = NULL
)
}
\arguments{
\item{obj}{Seurat object}

\item{features}{gene features}

\item{assay}{assay in obj}

\item{layer}{layer in assay}

\item{nameprefix}{prefix to metrics}
}
\value{
data frame
}
\description{
Metrics:
sum = number of expressed genes per cell.
rel = fraction of genes expressed per cell.
abs = total (summed) expression
norm = sum of rescaled \link{0,3} expression, max is 3*length(features), rounded
lognorm = log2 of norm (dapening extreme values)
Check out: UCell::AddModuleScore_UCell
Check out: scexpr::fgsea_groupwise
}
\examples{
\dontrun{
coexpr <- coexpression_metrics(obj = so,
                               features = c("NKG7", "PRF1", "KLRG1"),
                               nameprefix = "killer")
so <- Seurat::AddMetaData(so, coexpr)
}
}
