% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_UMI_matrices.R
\name{compare_UMI_matrices}
\alias{compare_UMI_matrices}
\title{Compare entries of two compatible UMI matrices}
\usage{
compare_UMI_matrices(mat1, mat2)
}
\arguments{
\item{mat1}{matrix 1 with rownames and colnames}

\item{mat2}{matrix 2 with rownames and colnames}
}
\value{
list with measures of differences or errors, as a whole and rowwise
}
\description{
Matrices will be reduces to matching rownames and columnnames.
Then root mean square error (RSME), total and relative differences in UMIs counts
are computed. This is done for matrix as a whole as well as rowwise. Rows are expected
to contain features. It is expected that matrices contain positive values only.
An error will not be thrown if this is violated, though.
}
\examples{
\dontrun{
mat1 <- random_matrix_normal <- matrix(rnorm(5 * 6), nrow = 5, ncol = 6)
mat2 <- random_matrix_normal <- matrix(rnorm(5 * 6), nrow = 5, ncol = 6)
colnames(mat1) <- colnames(mat2) <- 1:ncol(mat1)
rownames(mat1) <- rownames(mat2) <- 1:nrow(mat1)
mat_diff <- compare_UMI_matrices(mat1 = mat1, mat2 = mat2)
}
}
