% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_barplot.R
\name{composition_barplot}
\alias{composition_barplot}
\title{Plot composition of one category by another}
\usage{
composition_barplot(
  SO,
  x_cat,
  fill_cat,
  col_pal = scexpr::col_pal("custom"),
  border_color = "black",
  labels = F,
  min_label = 0.05,
  in_percent = T,
  label_decimals = 2
)
}
\arguments{
\item{SO}{Seurat object or data frame}

\item{col_pal}{}
}
\description{
Plot composition of one category by another
}
