% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_barplot.R
\name{composition_barplot}
\alias{composition_barplot}
\title{Plot composition of one category by another}
\usage{
composition_barplot(SO, x_cat, fill_cat, col_pal = scexpr::col_pal("custom"))
}
\arguments{
\item{fill_cat}{}
}
\value{

}
\description{
Plot composition of one category by another
}
