% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition_barplot.R
\name{composition_barplot}
\alias{composition_barplot}
\title{Plot composition of one category by another}
\usage{
composition_barplot(
  SO,
  x_cat,
  fill_cat,
  y = c("rel", "abs"),
  col_pal = scexpr::col_pal("custom"),
  border_color = "black",
  plot_rel_labels = F,
  label_only_largest = F,
  plot_abs_labels = F,
  plot_total_labels = F,
  total_labels_ypos = 1.05,
  plot_sigma = F,
  sigma_just = ifelse(flip, 2, -1),
  min_label_freq = 0.05,
  label_rel_pct = T,
  label_rel_pct_decimals = 2,
  summarize_all_x = F,
  label_rel_nudge = c(0, 0),
  label_abs_nudge = c(0, 0),
  label_size = 4,
  flip = F
)
}
\arguments{
\item{SO}{Seurat object or data frame}

\item{sigma_just}{}
}
\description{
Plot composition of one category by another
}
