% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_identifier.R
\name{convert_gene_identifier}
\alias{convert_gene_identifier}
\title{Title}
\usage{
convert_gene_identifier(
  gene_idents,
  ident_in = "SYMBOL",
  ident_out = "ENTREZID",
  species = c("Hs", "Mm")
)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Convert gene identifiers from ident_in to ident_out while taking care
of known duplicates and providing aliases.
}
\details{
keytypes(org.Hs.eg.db) or colums(org.Hs.eg.db) are: ACCNUM ALIAS ENSEMBL
ENSEMBLPROT ENSEMBLTRANS ENTREZID ENZYME EVIDENCE EVIDENCEALL
GENENAME GENETYPE GO GOALL IPI MAP OMIM ONTOLOGY ONTOLOGYALL
PATH PFAM PMID PROSITE REFSEQ SYMBOL UCSCKG UNIPROT
}
