% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_identifier.R
\name{convert_gene_identifier}
\alias{convert_gene_identifier}
\title{Title}
\usage{
convert_gene_identifier(
  gene_idents,
  ident_in = "SYMBOL",
  ident_out = "ENTREZID",
  species = c("Hs", "Mm")
)
}
\arguments{
\item{species}{}
}
\value{

}
\description{
Title
}
