% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_identifier.R
\name{convert_gene_identifier}
\alias{convert_gene_identifier}
\title{Title}
\usage{
convert_gene_identifier(
  idents,
  ident_in = "SYMBOL",
  ident_out = c("ENTREZID", "ALIAS", "GENENAME"),
  species = c("Hs", "Mm"),
  return = c("data.frame", "vector")
)
}
\arguments{
\item{idents}{character vector of gene indentifiers}

\item{ident_in}{type of identifier provided in idents}

\item{ident_out}{type(s) of identifier(s) to return}

\item{return}{}
}
\value{

}
\description{
Convert gene identifiers from ident_in to ident_out with AnnotationDbi while taking care of known duplicates and providing aliases.
For human org.Hs.eg.db is used, for mouse org.Mm.eg.db.
}
\details{
keytypes(org.Hs.eg.db) or colums(org.Hs.eg.db) are: ACCNUM ALIAS ENSEMBL
ENSEMBLPROT ENSEMBLTRANS ENTREZID ENZYME EVIDENCE EVIDENCEALL
GENENAME GENETYPE GO GOALL IPI MAP OMIM ONTOLOGY ONTOLOGYALL
PATH PFAM PMID PROSITE REFSEQ SYMBOL UCSCKG UNIPROT
}
