% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmat_to_hdf5.R
\name{countmat_to_hdf5}
\alias{countmat_to_hdf5}
\title{Save UMI counts from matrix to disk in initial 10X format}
\usage{
countmat_to_hdf5(
  mat,
  colsplit = NULL,
  out_dir,
  sub_dir = "filtered_feature_bc_matrix",
  types = c("sparse", "HDF5"),
  ...
)
}
\arguments{
\item{mat}{complete UMI count matrix}

\item{colsplit}{optionally: vector of factor levels to split mat by
(column wise), becomes separate folders in out_dir}

\item{out_dir}{output directory}

\item{sub_dir}{sub directory in out_dir or in folders of colsplit}

\item{types}{format(s) to save}

\item{...}{arguments to DropletUtils::write10xCounts}
}
\value{
nothing in R but files on disk
}
\description{
Save UMI counts from matrix to disk in initial 10X format
}
\examples{
\dontrun{
# obtain count_matrix from Seurat object or any other single cell format
# tcell_sub_split is a named list of data frames with one row per cell
# count_matrix is a combined mat of all cells in tcell_sub_split
countmat_to_hdf5(mat = count_matrix,
                 colsplit = rep(names(tcell_sub_split), purrr::map_int(tcell_sub_split, nrow)),
                 out_dir = "outdir",
                 types = "HDF5")
}
}
