% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_cluster_assignment_by_module_scores.R
\name{discretize_module_score_suggestive_clusters}
\alias{discretize_module_score_suggestive_clusters}
\title{Assign clusters based on multiple module scores}
\usage{
discretize_module_score_suggestive_clusters(
  obj,
  modules,
  method_score = c("module_score", "gsea"),
  method_cluster = c("louvain", "hclust"),
  n_times_modules_subcluster = 1,
  AddModuleScore_UCell_args = list(slot = "data"),
  fgseaMultilevel_args = list(pathways = modules),
  get_layer_args = list(),
  gsea_overcluster_resolution = 2
)
}
\arguments{
\item{obj}{seurat object}

\item{modules}{list of modules/gene sets/pathways}

\item{method_score}{scoring method}

\item{method_cluster}{cluster method}

\item{AddModuleScore_UCell_args}{args to UCell::AddModuleScore_UCell when
method_score is module_score}

\item{fgseaMultilevel_args}{args to fgsea::fgseaMultilevel when method_score
is gsea}

\item{get_layer_args}{args to scexpr::getLayer in scexpr::fgsea_groupwise
when method_score is gsea}

\item{gsea_overcluster_resolution}{}
}
\value{
list
}
\description{
Assign clusters based on multiple module scores
}
