% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_jitter.R
\name{.expr_jitter}
\alias{.expr_jitter}
\title{Title}
\usage{
.expr_jitter(
  d,
  feat,
  pt.size,
  geom1 = "jitter",
  geom2 = "violin",
  font.size = 14,
  filter.non.expr = F,
  plot.expr.freq = F,
  label.size = 4,
  ngc,
  pgc,
  ngn,
  pgn
)
}
\arguments{
\item{pgn}{}
}
\value{

}
\description{
Title
}
