% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_seurat.R
\name{downsample_seurat}
\alias{downsample_seurat}
\title{Downsample a Seurat object}
\usage{
downsample_seurat(obj, downsample = 1, group = NULL)
}
\arguments{
\item{obj}{A Seurat object.}

\item{downsample}{Either:
\itemize{
\item a proportion in (0, 1], or
\item the number of cells to retain (> 1).
}}

\item{group}{Optional metadata column used for stratified sampling.}
}
\value{
A downsampled Seurat object.
}
\description{
Downsample a Seurat object
}
