% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_correlation.R
\name{feature_correlation}
\alias{feature_correlation}
\title{Title}
\usage{
feature_correlation(
  SO,
  assay = c("RNA", "SCT"),
  method = c("pearson", "spearman", "kendall"),
  features,
  cells = NULL,
  min_pct = 0.1,
  limit_p = 1e-303,
  lm_resid = F,
  bar_fill = c("correlation_sign", "ref_feature_pct"),
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
