% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot.R
\name{feature_plot}
\alias{feature_plot}
\title{Title}
\usage{
feature_plot(
  SO,
  features,
  assay = c("RNA", "SCT"),
  dims = c(1, 2),
  cells = NULL,
  downsample = 1,
  make.cells.unique = F,
  pt.size = 1,
  pt.size.expr.factor = 1,
  order = T,
  min.q.cutoff = 0,
  max.q.cutoff = 1,
  reduction = "tsne",
  split.by = NULL,
  shape.by = NULL,
  combine = T,
  ncol.combine = NULL,
  nrow.combine = NULL,
  nrow.inner = NULL,
  ncol.inner = NULL,
  feature.aliases = NULL,
  binary = F,
  title = NULL,
  title.font.size = 14,
  cutoff.feature = NULL,
  cutoff.expression = 0,
  exclusion.feature = NULL,
  legend.position = "right",
  legend.title.text.size = 14,
  legend.text.size = 10,
  legend.barheight = 3,
  legend.barwidth = 0.5,
  legend.nrow = NULL,
  legend.ncol = NULL,
  legend.key.size = 0.3,
  legend.shape.size = 3,
  legend.col.size = 3,
  hide.shape.legend = F,
  font.family = "sans",
  col.pal.c = "spectral",
  col.pal.d = "custom",
  col.excluded.cells = "grey95",
  col.non.expresser = "grey85",
  col.expresser = "tomato2",
  col.pal.rev = F,
  theme = ggplot2::theme_bw(),
  plot.axis.labels = F,
  plot.panel.grid = F,
  plot.freq.title = T,
  plot.freq = T,
  plot.legend.title = F,
  plot.title = T,
  order.discrete = NULL,
  freq.position = c(-Inf, Inf),
  freq.font.size = 4,
  plot.cutoff.feature.inset = F,
  inset.position = c(0.75, 0.1),
  inset.size = c(0.15, 0.15),
  strip.font.size = 14,
  strip.selection = NA,
  plot.labels = NULL,
  label.size = 12,
  ...
)
}
\arguments{
\item{order}{for meta.col: remains T if var is continuous but becomes F if var is integer (~probably discrete)}

\item{label.size}{}
}
\value{

}
\description{
Title
}
