% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot_stat.R
\name{feature_plot_stat}
\alias{feature_plot_stat}
\title{Title}
\usage{
feature_plot_stat(
  SO,
  features,
  meta_col,
  assay = "RNA",
  geom1 = c("jitter", "point"),
  geom2 = c("boxplot", "violin", "none"),
  jitterwidth = 0.2,
  label.size = 4,
  plot.expr.freq = F,
  expr.freq.decimals = 2,
  expr.freq.pct = F,
  expr.freq.hjust = 0.3,
  plot.non.expr = T,
  plot.strip = T,
  cells = NULL,
  downsample = 1,
  legend.title = "SO.split",
  split.by = NULL,
  pt.size = 0.5,
  col.pal = "custom",
  col.pal.dir = 1,
  feature.aliases = NULL,
  feature_cut = NULL,
  feature_cut_expr = 0,
  feature_ex = NULL,
  theme = ggplot2::theme_bw(),
  facetting_args = list(scales = "free_y")
)
}
\arguments{
\item{plot.expr.freq}{to plot labels of expression frequencies}

\item{expr.freq.decimals}{decimal precision of frequency labels}

\item{expr.freq.pct}{plot expression frequency in percent?}

\item{plot.non.expr}{do plot dots of non-expressers (0 UMIs)}
}
\description{
Title
}
