% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_plot_stat.R
\name{feature_plot_stat}
\alias{feature_plot_stat}
\title{Title}
\usage{
feature_plot_stat(
  SO,
  features,
  assay = c("RNA", "SCT"),
  geom1 = c("jitter", "point", "dotplot"),
  geom2 = c("violin", "boxplot", "none"),
  jitterwidth = 0.2,
  label.size = 4,
  meta.col,
  meta.col.levels = NULL,
  plot.expr.freq = F,
  filter.non.expr = F,
  cells = NULL,
  downsample = 1,
  split.by = NULL,
  pt.size = 1,
  combine = T,
  ncol.combine = NULL,
  nrow.combine = NULL,
  nrow.inner = NULL,
  ncol.inner = NULL,
  feature.aliases = NULL,
  cutoff.feature = NULL,
  cutoff.expression = 0,
  exclusion.feature = NULL,
  plot.panel.grid = F,
  font.family = "sans",
  make.cells.unique = F,
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
