% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsgea_on_msigdbr.R
\name{fgsea_on_msigdbr}
\alias{fgsea_on_msigdbr}
\title{Title}
\usage{
fgsea_on_msigdbr(
  gene.ranks,
  use.msigdbr = T,
  gene.sets = NULL,
  min.padj = 0.001,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
