% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsgea_on_msigdbr.R
\name{fgsea_on_msigdbr}
\alias{fgsea_on_msigdbr}
\title{Convenient wrapper around fgsea function}
\usage{
fgsea_on_msigdbr(
  gene.ranks = NULL,
  gene.sets = NULL,
  return.subset.gene.sets = T,
  return.gene.sets = T,
  min.padj = 0,
  use.msigdbr = F,
  msigdbr_args = list(species = "Homo sapiens", collection = NULL, subcollection = NULL),
  fgsea_fun = fgsea::fgseaMultilevel,
  fgsea_args = list(),
  ...
)
}
\arguments{
\item{gene.ranks}{this becomes stats in fgsea_fun}

\item{gene.sets}{this becomes pathways in fgsea_fun}

\item{return.subset.gene.sets}{return the subset of each gene set that is actually found in gene.ranks}

\item{return.gene.sets}{set to FALSE in order to save memory when the same gene sets are used multiple times}

\item{...}{arguments to .plotEnrichment}
}
\description{
Convenient wrapper around fgsea function
}
\examples{
\dontrun{
# get gene sets from msigdbr, split to a named list
gene.sets <- scexpr:::.get.split.msigdbr()
msigdbr::msigdbr_collections()
# provide gene.sets manually and leave use.msigdbr = F in case fgsea_on_msigdbr is called many times with the same gene.sets
}
}
