% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_marker.R
\name{find_all_marker}
\alias{find_all_marker}
\title{Calculate markers with presto}
\usage{
find_all_marker(
  obj,
  meta_col,
  assay = "RNA",
  layer = "data",
  calc_avglog2fc = T,
  eps = 0.000001,
  features = NULL,
  mc.cores = 1
)
}
\arguments{
\item{obj}{seurat object}

\item{meta_col}{column name in meta.data}

\item{assay}{assay in obj}

\item{layer}{which layer to pull expression matrix from}

\item{calc_avglog2fc}{calculate avglog2fc like Seurat, prestos logFC is diff
of means (wrong naming)}

\item{eps}{pseudo count to avoid div by zero}

\item{features}{features selection to increase speed further}

\item{mc.cores}{multicores for calc_avglog2fc}
}
\value{
data frame with DEG
}
\description{
Column logFC is difference in group means, not logFC. But avg_log2FC
in Seurat style can be added.
}
