% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_in_expr_dich.R
\name{find_marker_in_expr_dich}
\alias{find_marker_in_expr_dich}
\title{Find marker genes in expressing cells; dichotomous}
\usage{
find_marker_in_expr_dich(obj, group, get_layer_args = list(), mc.cores = 4)
}
\arguments{
\item{obj}{seurat object or gene x cell matrix (cells as columns)}

\item{group}{meta column in object or group vector of length ncol(obj);
must have two levels}

\item{get_layer_args}{args to get_layer}

\item{mc.cores}{parallel computing: mc.cores}
}
\description{
Find marker genes in expressing cells; dichotomous
}
