% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_pairwise.R
\name{find_markers_pairwise}
\alias{find_markers_pairwise}
\title{Find marker genes in pairwise comparisons}
\usage{
find_markers_pairwise(obj, group, split = NULL, mc.cores = 1, ...)
}
\arguments{
\item{...}{}
}
\description{
Find marker genes in pairwise comparisons
}
