% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_expression.R
\name{freq_expression}
\alias{freq_expression}
\title{Title}
\usage{
freq_expression(
  obj,
  features,
  meta_col,
  assay = "RNA",
  cells = NULL,
  split_feature = NULL,
  format_pct = F
)
}
\description{
Title
}
