% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_pie_chart.R
\name{freq_pie_chart}
\alias{freq_pie_chart}
\title{Title}
\usage{
freq_pie_chart(
  SO,
  meta_col,
  order = T,
  fill = colrr::col_pal("custom"),
  color = "white",
  radius_inside = 0.3,
  label_outside = c("none", "abs", "rel"),
  label_inside = c("rel", "abs", "none"),
  label_rel_cutoff = 0,
  label_size = 5,
  label_radius_inside = 0.75,
  label_radius_outside = 1.1,
  label_angle_inside = NULL,
  label_angle_outside = NULL,
  label_overlap = c("ignore", "alternate", "outside"),
  overlap_outside_radius = 1.1,
  label_rel_pct = F,
  label_rel_dec = 2,
  legend_title = NULL,
  theme_args = list(panel.grid = ggplot2::element_blank(), axis.title =
    ggplot2::element_blank(), axis.text = ggplot2::element_blank(), axis.ticks =
    ggplot2::element_blank())
)
}
\arguments{
\item{SO}{Seurat object or data frame}

\item{meta_col}{column from meta data to use}
}
\description{
Title
}
