% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_pie_chart.R
\name{freq_pie_chart}
\alias{freq_pie_chart}
\title{Title}
\usage{
freq_pie_chart(
  SO,
  meta.col,
  inset.text.size = 5,
  inset.text.radius = 0.75,
  legend.position = "right",
  col_pal = scexpr::col_pal("custom"),
  order_pieces = T
)
}
\arguments{
\item{SO}{Seurat object or data frame}

\item{order_pieces}{}
}
\description{
Title
}
