% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data frame for plotting from Seurat object(s)}
\usage{
get_data(
  SO,
  feature,
  reduction = "umap",
  dims = c(1, 2),
  assay = "RNA",
  layer = "data",
  cells = NULL,
  meta_col = NULL,
  qmin = 0,
  qmax = 1,
  order = T,
  order_rev = F,
  order_abs = T,
  order_discr = T,
  shuffle = F,
  bury_NA = T,
  na_rm = F,
  inf_rm = F,
  label_feature = NULL,
  contour_feature = NULL,
  split_feature = NULL,
  shape_feature = NULL,
  trajectory_slot = NULL,
  downsample = 1,
  feature_cut = NULL,
  feature_cut_expr = 0,
  feature_ex = NULL
)
}
\arguments{
\item{SO}{one Seurat object or a list of multiple ones}

\item{feature}{vector of features to fetch (genes or column names in
meta data)}

\item{reduction}{reduction to fetch from reductions slot of SO}

\item{dims}{dimensions of the selected reduction to extract from SO}

\item{assay}{which assay to get expression data from}

\item{layer}{which layer/slot to get expression data from, e.g. counts or data}

\item{cells}{vector of cell names to select for regular plotting (=1),
deselected ones are plotted with color col_ex_cells (=0) in feature_plot}

\item{meta_col}{meta data columns to attach}

\item{qmin}{lower quantile of feature values where to cut the color scale}

\item{qmax}{upper quantile}

\item{order}{order rows from lowest to highest (negative at bottom)}

\item{order_rev}{reverse row order}

\item{order_abs}{absolute values for ordering: zeros first and anything
distant from zero on top}

\item{order_discr}{set TRUE to for order by abundance, only relevant for
discrete meta features}

\item{shuffle}{do shuffle rows (if order is FALSE) to randomize plotting
order; if !order and !shuffle the provided order is not altered}

\item{bury_NA}{put NA to the bottom which will bury them upon plotting}

\item{na_rm}{remove NA values? (only for meta feature)}

\item{inf_rm}{remove infinite values?}

\item{label_feature}{feature for plotting labels}

\item{contour_feature}{feature for plotting contours}

\item{split_feature}{feature for creating facets}

\item{shape_feature}{feature to shape plotted points by}

\item{downsample}{random downsample cells which will completely remove them
from returned data frame and hence influence statistics and plotting}

\item{feature_cut}{filter cells by a one or more cutoff features,
cells with UMI>feature_cut_expr are excluded and get col_ex_cells}

\item{feature_cut_expr}{numeric vector of cutoff expressions}

\item{feature_ex}{filter cells by an exclusion feature, cells with UMI>0
are excluded and get col_ex_cells as color, passed to scexpr::get_data}
}
\value{
a list of data frame
}
\description{
Get data frame for plotting from Seurat object(s)
}
