% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer.R
\name{get_layer}
\alias{get_layer}
\title{Get layer data with a few fallback mechanisms}
\usage{
get_layer(
  obj,
  assay = "RNA",
  layer = "data",
  features = NULL,
  cells = NULL,
  as = c("sparse", "dense", "df"),
  transpose = F
)
}
\arguments{
\item{obj}{Seurat object}

\item{assay}{assay}

\item{layer}{layer or slot}

\item{features}{filter for features}

\item{cells}{filter for cells}

\item{as}{what to return, sparse or dense matrix or data frame}

\item{transpose}{apply Matrix::t() ?}
}
\value{
matrix (sparse/dense) or data frame
}
\description{
Works with v4 and v5 objects. Join layers in v5 objects if possible
and necessary. Returns NULL if layer is 0x0 matrix.
}
\examples{
\dontrun{
get_layer(SO, assay = "SCT", layer = "counts", features = c("CD4", "CD8A"),
          as = "dense")
}
}
