% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea_on_msigdbr.R
\name{gsea_on_msigdbr}
\alias{gsea_on_msigdbr}
\title{Convenient wrapper around fgsea function}
\usage{
gsea_on_msigdbr(
  gene_ranks,
  gene_sets = NULL,
  return_gene_sets_subset = T,
  return_gene_sets = T,
  use_msigdbr = F,
  msigdbr_args = list(db_species = "HS", species = "human", collection = c("C1", "C2",
    "C3", "C4", "C5", "C6", "C7", "C8", "H")),
  fgsea_fun = fgsea::fgseaMultilevel,
  fgsea_args = list(),
  ...
)
}
\arguments{
\item{gene_ranks}{this becomes stats in fgsea_fun}

\item{gene_sets}{this becomes pathways in fgsea_fun}

\item{return_gene_sets_subset}{return the subset of each gene set that is actually found in gene_ranks}

\item{return_gene_sets}{set to FALSE in order to save memory when the same gene sets are used multiple times}

\item{...}{arguments to .plotEnrichment}
}
\description{
Convenient wrapper around fgsea function
}
\examples{
\dontrun{
#' # H: hallmark gene sets
# C1: positional gene sets
# C2: curated gene sets
# C3: regulatory target gene sets
# C4: computational gene sets
# C5: ontology gene sets
# C6: oncogenic signature gene sets
# C7: immunologic signature gene sets
# C8: cell type signature gene sets
s2ntab <- scexpr::gsea_s2n_groupwise(obj = so, group = "integrated_snn_res.0.7")
msiggsea <- gsea_on_msigdbr(gene_ranks = s2ntab[,"20"], use_msigdbr = T)
# get gene sets from msigdbr, split to a named list
gene_sets <- scexpr:::.get.split.msigdbr()
msigdbr::msigdbr_collections()
# provide gene_sets manually and leave use_msigdbr = F in case gsea_on_msigdbr is called many times with the same gene_sets
}
}
