% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pseudobulk.R
\name{heatmap_pseudobulk}
\alias{heatmap_pseudobulk}
\title{Title}
\usage{
heatmap_pseudobulk(
  SO,
  assay = c("RNA", "SCT"),
  meta.col,
  meta.col.levels,
  meta.col.levels.select,
  features = NULL,
  feature.labels,
  topn.features = 1e+06,
  topn.metric = c("logFC", "auc", "padj"),
  min.pct = 0.1,
  max.padj = 0.05,
  title = NULL,
  title.font.size = 14,
  y.font.size = 10,
  legend.position = "right",
  legend.direction = "vertical",
  tile.borders = T,
  dotplot = F,
  plot.feature.breaks = T,
  plot.secondary.axis.with.full.genenames = F,
  legend.title.text.size = 10,
  legend.barheight = 8,
  legend.barwidth = 1,
  legend.text.size = 10,
  legend.title = "z-score of avg expr",
  legend.title.position = "top",
  feature.labels.nudge_x = -0.1,
  feature.labels.axis.width = 0.2,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
