% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pseudobulk.R
\name{heatmap_pseudobulk}
\alias{heatmap_pseudobulk}
\title{Title}
\usage{
heatmap_pseudobulk(
  SO,
  meta.col = NULL,
  levels.calc = NULL,
  levels.plot = NULL,
  assay = c("RNA", "SCT"),
  features = NULL,
  order_features = F,
  normalization = "scale",
  topn.features = 10,
  topn.ties = F,
  topn.metric = c("padj", "auc", "logFC"),
  min.pct = 0.1,
  max.padj = 0.05,
  title = NULL,
  title.font.size = 14,
  font.family = "sans",
  y.font.size = 10,
  tile.borders = T,
  dotplot = F,
  plot.feature.breaks = T,
  plot.sec.axis = F,
  legend.position = "right",
  legend.direction = "vertical",
  legend.barheight = 8,
  legend.barwidth = 1,
  legend.text.size = 10,
  legend.title.text.size = 10,
  legend.title.fill = NULL,
  legend.title.size = NULL,
  legend.title.position = "top",
  legend.labels = c("min", "", "max"),
  feature.labels = NULL,
  feature.labels.nudge_x = -0.1,
  feature.labels.axis.width = 0.2,
  plot_hlines_between_groups = F,
  hlines = NULL,
  hlines_args = list(),
  ...
)
}
\arguments{
\item{SO}{Seurat object}

\item{meta.col}{which column from meta.data of SO to use as x-axis; if NULL current Idents(SO) are used}

\item{levels.calc}{which levels in meta.col to include in calculation;
all levels if NULL; the order provided defines order on x-axis; level selection
will affect scaling calculations}

\item{levels.plot}{which levels in meta.col to include in platting;
if NULL this equals to levels.calc; must be subset of levels.calc;
the order provided defines order on x-axis; this will not affect scaling
calculation but only select levels for plotting}

\item{assay}{which assay to obtain expression values from}

\item{features}{optionally choose which features to plot (supervised)}

\item{order_features}{when features are provided should they be ordered automatically to generate a pretty heatmap;
not relevant when is.null(features)}

\item{normalization}{how to scale expression values; may be "scale" to
use base::scale and transform average expression value to a standardized
normal distribution or may be a numeric vector of length 2 for scaling
each feature from min (first value) to max (second value); in the latter case
c(-1,1) is most meaningful}

\item{topn.features}{if no features are selected, this will select how many
features to plot per level in meta.col; selection is done based on the metric
selected in topn.metric; respective features with greatest difference between
meta.col levels are selected (best DE features so to say)}

\item{topn.ties}{passed as 'with_ties' to dplyr::slice_min or dplyr::slice_max;
if set to FALSE exactly topn.features are returned even if there are rank-ties}

\item{topn.metric}{which differential expression metric to apply for feature
selection}

\item{min.pct}{filter features across levels in meta.col for a minimal fraction
of expressing cells; 0.1 stands for min 10 \% expressing cells}

\item{max.padj}{filter features across levels in meta.col for a significance
level of differential expression (one level vs. all others)}

\item{title}{which title to add to plot; if NULL no title is plotted}

\item{title.font.size}{font size of title}

\item{font.family}{which font type (family) to use for plotting,
e.g. mono or sans}

\item{y.font.size}{font size of features names on y-axis}

\item{tile.borders}{plot black tile border (TRUE) or not (FALSE);
better set to FALSE for may features and optionally set to TRUE for few
features only}

\item{dotplot}{plot heatmap as a dotplot which will also reveal the fraction
of expressing cells by dot sizes}

\item{plot_hlines_between_groups}{plot horizontal lines to separate groups of genes that are
differentially expressed by a group on the x-axis; if hlines is NULL the breaks are
determined automatically}

\item{hlines}{provide manual breaks of where to plot horizontal lines;
even though the y-axis is discrete, a numeric vector works here:
hlines = c(1.5, 3.5) will plot lines between the first and second, and, third and fourth
entry on the y-axis, respectively}

\item{hlines_args}{arguments to geom_hline like linewidth or linetype}
}
\description{
Title
}
