% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_pseudobulk2.R
\name{heatmap_pseudobulk2}
\alias{heatmap_pseudobulk2}
\title{Plot a heatmap of average gene transcription of transcriptomes by
different groups (clusters)}
\usage{
heatmap_pseudobulk2(
  SO,
  meta_col = NULL,
  levels_calc = NULL,
  levels_plot = NULL,
  assay = "RNA",
  min_pct = 0.1,
  max_padj = 0.05,
  features = NULL,
  feature_order = c("custom", "hclust", "none"),
  group_order = c("hclust", "none", "custom"),
  topn_metric = c("padj", "auc", "logFC"),
  topn_ties = F,
  dotplot = F,
  dotsize_range = c(2, 7),
  fill = "..auto..",
  color = "..auto..",
  scale = c("zscore", "1", "none"),
  features_topn = NULL,
  featurelabels = "..auto..",
  featurelabels_repel = F,
  featuresitalic = T,
  color_linewidth = 0.2,
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  colorsteps = "..auto..",
  colorsteps_nice = T,
  axes_flip = F,
  group_seplines = F,
  seplines_args = list(),
  theme = colrr::theme_material(),
  legend_fill_args = list(label.theme = ggplot2::element_text(size = 10), title.theme =
    ggplot2::element_text(size = 10), title.position = "top", title = "..auto..",
    title.hjust = 0.5, barwidth = 1, barheight = 8, order = 1),
  legend_size_args = list(label.theme = ggplot2::element_text(size = 10), title.theme =
    ggplot2::element_text(size = 10), title.position = "top", title =
    "transcription\\nfrequency [\%]", title.hjust = 0.5, label.position = "bottom", order
    = 2, ncol = NULL, nrow = NULL, override.aes = list(color = "white")),
  theme_args = list(axis.title = ggplot2::element_blank(), panel.grid =
    ggplot2::element_blank(), axis.text.x = ggplot2::element_text(angle = 40, hjust = 1),
    legend.position = "right", legend.direction = "vertical", legend.box = "horizontal"),
  repel_args = list(featurelabels_width = 0.2, featurelabels_nudhe_x = -1),
  sec_axis = F,
  convert_gene_identifier_args = list(ident_in = "SYMBOL", ident_out = "GENENAME")
)
}
\arguments{
\item{SO}{Seurat object}

\item{meta_col}{which column from meta.data of SO to use as x-axis;
if NULL current Idents(SO) are used}

\item{levels_calc}{which levels in meta_col to include in calculation;
all levels if NULL; the order provided defines order on x-axis; thisn level
selection will affect scaling calculations}

\item{levels_plot}{which levels in meta_col to include in platting;
if NULL this equals to levels_calc; must be subset of levels_calc;
the order provided defines order on x-axis; this will not affect scaling
calculation but only select levels for plotting;
ordering will not work perfectly when more than 1 SO is provided and factor levels
in meta_cols are not unique}

\item{assay}{which assay to obtain expression values from}

\item{min_pct}{for marker feature selection per only: min fraction of expressing
cells for a gene to become marker gene for a group}

\item{max_padj}{for marker feature selection per only: max adjusted
p-value for a gene to become marker gene}

\item{features}{choose which features to plot; if NULL features_topn becomes
relevant}

\item{feature_order}{if and how to order features?}

\item{group_order}{if and how to order groups?}

\item{topn_metric}{which differential expression metric to apply for feature
selection; only relevant if features_topn is not NULL}

\item{topn_ties}{break ties for features_topn? if TRUE, more then
features_topn may be plotted}

\item{dotplot}{plot dots instead of tiles with percent of expressing cells
as dot size}

\item{dotsize_range}{range for dot size}

\item{fill}{color palette vector for fill of tiles or dots. when auto,
RColorBrewer::RdBu is used}

\item{color}{color of stroke (border) around tiles or dots; "NA" means no
stroke is plotted; NA has
to be put in quotation mark ("NA"), such that geom_point accepts it.
other choices may be black, white or any other color code; when "auto",
grey70 is used by default when is.null(dotsizes) and a the number of
features is below 100.}

\item{scale}{how to scale values: not, zscore or from -1 to 1}

\item{features_topn}{only plot the top n features (ordered by value) per
group}

\item{featurelabels}{subset of feature labels to plot; NULL to plot all,
"" to plot none; can be a named vector with names being aliases to use for
plotting e.g. c("CD20" = "MS4A1", "CD3", "KLRG1") to only alter MS4A1;
auto to omit labels by default when more than 100 features are there}

\item{featurelabels_repel}{do repel feature axis labels?}

\item{featuresitalic}{shorthand to plot feature labels in italic}

\item{color_linewidth}{linewidth of borders (stroke) around tiles or dots}

\item{legendbreaks}{a single number, a vector of explicit breaks, or "auto"
for ggplot default or "minmidmax" for three breaks at minimum, middle and
maximum of value range}

\item{legendlabels}{labels for breaks, e.g. c("min", "mid", "max")}

\item{colorsteps}{NULL to have normal colorbar, auto for default colorsteps,
a single number or a vector of explicit steps; may not work with any number
when colorsteps_nice is TRUE}

\item{colorsteps_nice}{heuristic for pretty steps}

\item{axes_flip}{do flip them?}

\item{group_seplines}{plot lines that separate features belonging to
different groups}

\item{seplines_args}{arguments to geom_hline for seplines}

\item{theme}{ggplot2 theme}

\item{legend_fill_args}{arguments to ggplot2::guide_colorsteps or
ggplot2::guide_colorbar, depend upon the color scale}

\item{legend_size_args}{arguments to ggplot2::guide_legend to modify the
size legend; e.g. use override.aes = list(size = c(1,3,5)) to adjust dot size
in legend in contrast to dotsize_range, one number for each dot in legend
needed}

\item{theme_args}{arguments to ggplot2::theme}

\item{repel_args}{fine tuning for featurelabel repelling}

\item{sec_axis}{plot a secondary feature axis with full gene names}

\item{convert_gene_identifier_args}{arguments to
scexpr::convert_gene_identifier}
}
\description{
Plot a heatmap of average gene transcription of transcriptomes by
different groups (clusters)
}
