% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hla_df_from_xml.R
\name{hla_df_from_xml}
\alias{hla_df_from_xml}
\title{Prepare a data frame from a xml file containing hla allele information}
\usage{
hla_df_from_xml(file_path, lapply_fun = lapply, ...)
}
\arguments{
\item{file_path}{xml file path}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{
a data frame
}
\description{
The xml file required for this function may be downloaded from the GitHub repository of the IMGT/HLA organization:
https://github.com/ANHIG/IMGTHLA/blob/Latest/xml/hla.xml.zip. The function then creates a datafrema from this xml
which can be used subsequently to match reads of HLA-loci from bulk or single cell RNA seq.
}
\details{
HLA nomenclature: http://hla.alleles.org/nomenclature/naming.html
}
\examples{
\dontrun{
hla_df_from_xml(file_path = "path/hla.xml", lapply_fun = parallel::mclapply,
mc.cores = parallel::detectCores())
}
}
