% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeltransfer_module_gsea.R
\name{labeltransfer_module_gsea}
\alias{labeltransfer_module_gsea}
\title{Transfer cluster labels, i.e. celltype classification based on gene sets or
modules}
\usage{
labeltransfer_module_gsea(
  test_obj,
  test_clusters,
  modules,
  AddModuleScore_UCell_args = list(ncores = 1, name = "")
)
}
\arguments{
\item{test_obj}{Seurat object}

\item{test_clusters}{meta.data column in test_obj}

\item{modules}{list of named gene sets / modules / pathways}

\item{AddModuleScore_UCell_args}{arguments to UCell::AddModuleScore_UCell}
}
\value{
list
}
\description{
In comparison to labeltransfer_singler a test object and modules (gene sets)
are stater material but not a ref object. Modules may be DE genes defining a
cell population in another dataset. Module score and test for enrichment with
fgsea are used to find matching populations (clusters) in test_obj.
}
