% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_objects.R
\name{merge_objects}
\alias{merge_objects}
\title{Merge list of Seurat objects}
\usage{
merge_objects(obj_list, obj_ident_col = "obj_ident")
}
\arguments{
\item{obj_list}{list of Seurat objects.}

\item{obj_ident_col}{new column name to identify origin of cells by object}
}
\description{
Currently counts of RNA assay and meta data only.
}
