% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_objects.R
\name{merge_objects}
\alias{merge_objects}
\title{Merge list of Seurat objects}
\usage{
merge_objects(
  obj_list,
  obj_ident_col = "obj_ident",
  merge_reductions = F,
  misc_from_first = F
)
}
\arguments{
\item{obj_list}{list of Seurat objects.}

\item{obj_ident_col}{new column name to identify origin of cells by object}

\item{merge_reductions}{try to merge reductions?}

\item{misc_from_first}{use misc slot from first obj in list?}
}
\value{
merged seurat
}
\description{
Currently counts of RNA assay and meta data only.
}
\examples{
\dontrun{
so <- merge_objects(obj_list = list(so1,so2,so3))
}
}
