% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gsea.R
\name{plot_gsea}
\alias{plot_gsea}
\title{Plot results from scexpr::fgsea_on_msigdbr}
\usage{
plot_gsea(
  data,
  ticksSize = 0.2,
  color_ES_line = "black",
  color_min_max_ES_line = "black",
  color_leadingEdge = "hotpink2",
  label_genes = NULL,
  theme = ggplot2::theme_bw(),
  plot_leadingEdge_rank = T,
  plot_leadingEdge_size = T,
  annotation_pos = NULL,
  annotation_size = 4,
  rank_metric_zscore_lims = c(-4, 4),
  annotation_with_name = F
)
}
\arguments{
\item{annotation_with_name}{}
}
\description{
Plot results from scexpr::fgsea_on_msigdbr
}
