% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_SO.R
\name{prep_SO}
\alias{prep_SO}
\title{Prepare a somewhat standardized Seurat Object for further analyses}
\usage{
prep_SO(
  SO_unprocessed,
  samples = NULL,
  cells = NULL,
  min_cells = 50,
  downsample = 1,
  export_prefix = NULL,
  cluster_resolutions = 0.8,
  reductions = c("umap", "som", "gqtsom", "tsne"),
  nhvf = 800,
  npcs = 20,
  nintdims = 30,
  normalization = c("SCT", "LogNormalize"),
  batch_corr = c("harmony", "integration", "regression", "none"),
  ref_sample = NULL,
  integr_reduction = c("rpca", "cca", "rlsi"),
  vars.to.regress = NULL,
  seeed = 42,
  save_path = NULL,
  celltype_refs = NULL,
  celltype_label = "label.main",
  celltype_ref_clusters = NULL,
  diet_seurat = T,
  var_feature_filter = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{SO_unprocessed}{named list of split Seurat objects}

\item{samples}{names of SO_unprocessed that are to include; if missing all are used}

\item{cells}{vector of cell names to include; if missint all cells are used}

\item{min_cells}{min number of cells per object in SO_unprocessed; objects with lower cell number are excluded}

\item{downsample}{downsample a fraction of cells from each object in SO_unprocessed (downsample < 0);
or an absolute number of cells per object (downsample > 1 & downsample >= min_cells)}

\item{export_prefix}{prefix to the output rds file}

\item{cluster_resolutions}{vector of resolutions to compute by Seurat::FindClusters}

\item{reductions}{which reduction to calculate, tsne --> Seurat::RunTSNE, umap --> Seurat::RunUMAP}

\item{nhvf}{number high variables features, passed to Seurat::SCTransform
or Seurat::FindVariableFeatures}

\item{npcs}{number of principle components in calculate in pca and to
consider for downstream functions like tSNE or UMAP}

\item{nintdims}{number of integration dimension, when batch_corr = "integration"}

\item{normalization}{algorithm for normalization of UMIs}

\item{batch_corr}{algorithm for batch correction between samples in SO_unprocessed;
only relevant if more than 1 sample is passed}

\item{ref_sample}{optional names of the reference sample in SO_unprocessed for
Seurat::FindIntegrationAnchors or RunHarmony; reference; if NULL no reference is used}

\item{integr_reduction}{reduction method for Seurat::FindIntegrationAnchors; rpca is quicker than cca}

\item{vars.to.regress}{passed to Seurat::SCTransform or Seurat::ScaleData, applied if batch_corr == "regression"}

\item{seeed}{seed passed to several methods}

\item{save_path}{folder to save the resulting Seurat object as rds-file to}

\item{celltype_refs}{list(prim_cell_atlas = celldex::HumanPrimaryCellAtlasData(), MonacoImmune = celldex::MonacoImmuneData())}

\item{celltype_label}{list of label names to use from the reference data set (one list entry per celltype_refs; each entry may contain a vector of labels)}

\item{celltype_ref_clusters}{use a clustering as basis for grouped celltype annotation with SingleR. This will
fundamentally speed up the calculation!
e.g. if SCT assay is used and cluster_resolutions includes 0.8 then pass: SCT_snn_res.0.8.
when integration procedure assay is used: integrated_snn_res.0.8. For RNA assay: RNA_snn_res.0.8.}

\item{diet_seurat}{logical whether to run Seurat::DietSeurat}

\item{var_feature_filter}{character vector of features which are to exclude from variable features;
this will affect downstream PCA, dimension reduction and clustering;
e.g.: var_feature_filter = grep("^TR\link{ABGD}V", rownames(SOqc_split[\link{1}]), value = T)}

\item{verbose}{print messages and progress bars from functions}

\item{...}{additional arguments passed to harmony::RunHarmony, Seurat::RunUMAP,
Seurat::RunTSNE, Seurat::FindNeighbors, Seurat::FindClusters,
Seurat::PrepSCTIntegration, Seurat::FindIntegrationAnchors, Seurat::IntegrateData;
prefix arguments for RunHarmony by "RunHarmony__"; so e.g. RunHarmony__theta, prefix arguments
for EmbedSOM::SOM with "SOM__" (e.g. SOM__batch = T or SOM__rlen = 20,), for EmbedSOM::GQTSOM with "GQTSOM__" (e.g. GQTSOM__distf = 4)
and for EmbedSOM::EmbedSOM with "EmbedSOM__"}
}
\value{
Seurat Object, as R object and saved to disk as rds file
}
\description{
Prepare a somewhat standardized Seurat Object for further analyses
}
\examples{
\dontrun{
}
}
