% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_diagnostics.R
\name{qc_diagnostic}
\alias{qc_diagnostic}
\title{Title}
\usage{
qc_diagnostic(
  data.dir,
  nhvf = 1000,
  npcs = 10,
  min_nCount_RNA = 300,
  resolution = 0.8,
  SoupX = F,
  decontX = F,
  SoupX.resolution = 0.6,
  cells = NULL,
  invert_cells = F,
  qc_meta_resolution = 0.8,
  n_PCs_to_meta_clustering = 2,
  ...
)
}
\arguments{
\item{data.dir}{list or vector of full paths to filtered_feature_bc_matrix and raw_feature_bc_matrix, or filtered_feature_bc_matrix only (prohibiting the use of SoupX);
if not named, then the name of the common parent folder is uses as name(s)}

\item{SoupX}{logical whether to run SoupX. If T, raw_feature_bc_matrix is needed.}

\item{decontX}{logical whether to run celda::decontX to estimate RNA soup (contaminating ambient RNA molecules)}

\item{cells}{select to include}

\item{invert_cells}{invert cell selection, if TRUE cells are excluded}

\item{qc_meta_resolution}{the resolution (louvain algorithm) for clustering based on qc meta data and optionally additional PC dimensions (n_PCs_to_meta_clustering)}

\item{n_PCs_to_meta_clustering}{how many principle compoments (PCs) from phenotypic clustering to add to qc meta data;
this will generate a mixed clustering (PCs from phenotypes (RNA) and qc meta data like pct mt and nCount_RNA); the more PCs are added the greater the
phenotypic influence becomes}
}
\value{

}
\description{
On error in scDblFinder: Increase nhvf or try to change any other parameter
SoupX and/or decontX to filter cells or for correction
}
