% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_diagnostics.R
\name{qc_diagnostic}
\alias{qc_diagnostic}
\title{Title}
\usage{
qc_diagnostic(
  data.dir,
  nhvf = 500,
  npcs = 10,
  min_nCount_RNA = 300,
  resolutions = seq(0.6, 0.9, 0.1),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
