% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_params_meta_cols.R
\name{qc_params_meta_cols}
\alias{qc_params_meta_cols}
\title{Title}
\usage{
qc_params_meta_cols(
  SO,
  qc.cols = c("nFeature_RNA", "nCount_RNA", "pct_mt"),
  meta.cols,
  theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
