% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_params_meta_cols.R
\name{qc_params_meta_cols}
\alias{qc_params_meta_cols}
\title{Title}
\usage{
qc_params_meta_cols(
  SO,
  cells = NULL,
  meta_cols,
  qc_cols = c("nFeature_RNA", "nCount_RNA", "pct_mt"),
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
