% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{qc_plot2}
\alias{qc_plot2}
\title{Title}
\usage{
qc_plot2(
  SO,
  qc_cols = c("nCount_RNA_log", "nFeature_RNA_log", "pct_mt_log", "dbl_score_log",
    "residuals"),
  clustering_cols = c("RNA_snn_res.0.8", "meta_res.0.8_PC2"),
  reduction = "umapmetaPC2",
  geom2 = "boxplot"
)
}
\arguments{
\item{geom2}{}
}
\description{
Title
}
