% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{qc_plot2}
\alias{qc_plot2}
\title{Title}
\usage{
qc_plot2(
  SO,
  qc_cols = c("nCount_RNA_log", "nFeature_RNA_log", "pct_mt_log", "dbl_score_log",
    "pct_soup_decontX", "pct_soup_SoupX"),
  geom2 = "boxplot",
  save_path = NULL
)
}
\arguments{
\item{save_path}{}
}
\description{
Title
}
