% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot.R
\name{qc_plot_lin}
\alias{qc_plot_lin}
\title{Quickly plot common qc metrics}
\usage{
qc_plot_lin(
  SO1,
  SO2 = NULL,
  qc_cols = c("nCount_RNA", "nFeature_RNA", "pct_mt", "dbl_score"),
  x_cat = "orig.ident"
)
}
\arguments{
\item{SO1}{Seurat object 1}

\item{SO2}{Seurat object 2, optional}

\item{qc_cols}{QC columns on y-axes}

\item{x_cat}{meta data column for x-axis}

\item{sec_axis_lin}{plot a secondary axis with linear transformation of qc cols}
}
\description{
Plot qc metric, optional with a linear secondary axis.
When SO1 only is provided: Only plot cells from one Seurat object.
When SO1 and SO2 are provided:
Plot additionally those cells that have been filtered from that object with more cells.
Orig.
}
