% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_lognorm.R
\name{reverse_lognorm}
\alias{reverse_lognorm}
\title{Recreate counts layer from data layer}
\usage{
reverse_lognorm(
  SO,
  assay = "RNA",
  nCount_RNA = Seurat::Misc(SO, slot = "RNA_count_colSums"),
  scale.factor = 10000
)
}
\arguments{
\item{SO}{Seurat object}

\item{assay}{assay to use}

\item{nCount_RNA}{vector of colsums of RNA counts (lib sizes)}

\item{scale.factor}{factor as in Seurat::LogNormalize}
}
\value{
counts matrix to write to assay slot
}
\description{
Remove counts layer to save memory. Recreate it when needed.
}
\examples{
\dontrun{
  ## lognorm forward:
csums <- Matrix::colSums(counts)
xx <- sweep(counts, 2, csums, "/")
xx <- xx*1e4
data <- log1p(xx)

## SCT reverse:
# simply expm1 from data slot, see ?Seurat::SCTransform

# compare count matrices
rmse <- scexpr:::rmse_sparse(count_re, counts_orig)
equal <- all.equal(count_re, counts_orig)
}
}
