% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lda.R
\name{run_lda}
\alias{run_lda}
\title{Run linear discriminant analysis (LDA) for optimal separation of groups}
\usage{
run_lda(
  obj,
  groups,
  assay = "RNA",
  layer = "data",
  overwrite = F,
  features = Seurat::VariableFeatures(obj),
  reduction.name = "LDA",
  reduction.key = paste0(reduction.name, "_")
)
}
\arguments{
\item{obj}{seurat object}

\item{groups}{column with grouping in obj@meta.data}

\item{assay}{which assay to pull data from}

\item{layer}{which layer}

\item{overwrite}{overwrite existing dim red with sme name?}

\item{features}{which genes; less of them inreases calculation speed}

\item{reduction.name}{name of reduction to be created}

\item{reduction.key}{key of reduction to be created, needs trailing underscore}
}
\description{
Alternative to PCA. E.g. run clustering first, then LDA, then tsne or umap on
LDA. Or separate celltypes as labelled by SingleR in optimal way.
What does Seurat::RunLDA do ?
}
