% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2n_groupwise.R
\name{s2n_groupwise}
\alias{s2n_groupwise}
\title{Compute S2N per gene for each group (cluster) (one-vs-rest)}
\usage{
s2n_groupwise(obj, group, eps = 0.000001, ...)
}
\arguments{
\item{obj}{Seurat object or gene x cell matrix}

\item{group}{grouping column in obj meta.data or vector of length ncol(obj)}

\item{eps}{epsilon to avoid division by 0 in S2N formula}

\item{...}{arguments to scexpr::get_layer}
}
\value{
matrix
}
\description{
S2N(g, k) = (mean_k - mean_rest) / (sd_k + sd_rest + eps)
}
