% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO_prep04.R
\name{SO_prep04}
\alias{SO_prep04}
\title{Process assay of Seurat object}
\usage{
SO_prep04(
  SO,
  assay = "RNA",
  reductions = c("umap"),
  nhvf = 800,
  npcs = 20,
  normalization = c("SCT", "LogNormalize"),
  batch_corr = c("harmony", "none"),
  vars.to.regress = NULL,
  seeed = 42,
  var_feature_filter = NULL,
  verbose = F,
  FindVariableFeatures_args = list(),
  SCtransform_args = list(vst.flavor = "v2", method = "glmGamPoi", conserve.memory = F),
  RunPCA_args = list(),
  RunUMAP_args = list(),
  RunTSNE_args = list(theta = 0.01),
  FindNeighbors_args = list(dims = 1:npcs),
  FindClusters_args = list(resolution = seq(0.1, 0.8, 0.1)),
  RunHarmony_args = list(group.by.vars = "orig.ident"),
  interactive_varfeat_selection = F,
  interactive_varfeat_selection_inds = seq(max(nhvf/10, 50), min(3 * nhvf, nrow(SO)),
    length.out = 11),
  interactive_pc_selection = F,
  recalculate = F,
  ...
)
}
\arguments{
\item{SO}{seurat object}

\item{assay}{which assay to process}

\item{reductions}{which reduction to calculate, tsne and/or umap}

\item{nhvf}{number high variables features, passed to Seurat::SCTransform
or Seurat::FindVariableFeatures or Seurat::SelectIntegrationFeatures}

\item{npcs}{number of principle components in calculate in pca and to
consider for downstream functions like tSNE or UMAP}

\item{normalization}{algorithm for normalization of UMIs}

\item{batch_corr}{procedure for batch correction}

\item{vars.to.regress}{passed to Seurat::SCTransform or Seurat::ScaleData; will be applied
independent of what is set for batch_corr; if batch_corr is set to 'none', then only vars.to.regress is
used to regress out a variable in meta.data which may be sample-specific; other then that
it may not be meaningful to regress a sample-specific variable and perform batch_corr;
rather vars.to.regress may be used in combination with batch_corr to regress percent_mito or so}

\item{seeed}{random seed}

\item{var_feature_filter}{character vector of features which are to exclude from variable features;
this will affect downstream PCA, dimension reduction and clustering;
e.g.: var_feature_filter = \code{grep("^TR[ABGD]V", rownames(SOqc_split[[1]]), value = T)} to
exclude T cell receptor gene segments}

\item{verbose}{print messages and progress bars from functions}

\item{FindVariableFeatures_args}{arguments to Seurat::FindVariableFeatures}

\item{SCtransform_args}{arguments to Seurat::SCTransform but c("object", "assay", "new.assay.name", "seed.use", "verbose")}

\item{RunPCA_args}{arguments to Seurat::RunPCA}

\item{RunUMAP_args}{arguments to Seurat::RunUMAP}

\item{RunTSNE_args}{arguments to scexpr::run_fft_tsne}

\item{FindNeighbors_args}{arguments to Seurat::FindNeighbors}

\item{FindClusters_args}{arguments to Seurat::FindClusters}

\item{RunHarmony_args}{arguments to harmony::RunHarmony}

\item{interactive_varfeat_selection}{get asked?}

\item{interactive_varfeat_selection_inds}{get asked?}

\item{interactive_pc_selection}{do conduct interactive PC selection?}

\item{...}{not used}
}
\value{
seurat object
}
\description{
normalize, hvf, pca, batch corr, dim red, clustering
}
