% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano01_calc.R
\name{volcano01_calc}
\alias{volcano01_calc}
\title{Make a volcano plot comparing the gene expression of selected groups of cells from one or more preprocessed Seurat objects}
\usage{
volcano01_calc(
  SO,
  assay = "RNA",
  layer = "data",
  neg_cells,
  pos_cells,
  neg_name = "negative.group",
  pos_name = "positive.group",
  method = c("MAST", "wilcox", "wilcox_limma", "bimod", "roc", "t", "negbinom",
    "poisson", "LR", "DESeq2", "limma", "default"),
  fc_inf_squish = 2,
  p_zero_squish = 1,
  min_pct = 0.1,
  p_adj = c("bonferroni", "holm", "hochberg", "hommel", "BH", "BY", "fdr", "none"),
  fc_thresh = 0,
  feature_plot_args = list(col_pal_d_args = list(name =
    stats::setNames(colrr::col_pal("custom")[2:3], c(neg_name, pos_name)),
    missing_fct_to_na = F))
)
}
\arguments{
\item{SO}{one or more Seurat object(s); provide multiple objects as a (named) list}

\item{assay}{which assay to get expression data from}

\item{neg_cells}{vector of cell names for the negative group of cells; genes which
are expressed at higher level in this group will have a negative sign (minus); use Seurat::Cells()
or Seurat::WhichCells() to select or use filter operations on SO@meta.data to select cells}

\item{pos_cells}{vector of cell names for the positive group of cells; genes which
are expressed at higher level in this group will have a positive sign (plus); use Seurat::Cells()
or Seurat::WhichCells() to select or use filter operations on SO@meta.data to select cells}

\item{neg_name}{name for the negative group of cells; this will appear on plot axes}

\item{pos_name}{name for the positive group of cells; this will appear on plot axes}

\item{method}{backend for DE calculation;limma:intended for subsequent use MetaVolcanoR
which relies on values returned from limma; MAST: Seurat::FindMarkers;
with MAST as method; default: matrixTests::row_wilcoxon_twosample}

\item{fc_inf_squish}{numeric; genes with an infinite fold-change (happens when a gene not expressed at all in one group)
are not left infinite but will be put at: "largest finite fold-change +/- fc_inf_squish"; such genes are indicated
specifically}

\item{min_pct}{numeric; only consider genes for the volcano plot which are expressed at least by this
fraction of cells at least one group (neg_cells or pos_cells)}

\item{p_adj}{method for p-value adjustment}
}
\description{
Make a volcano plot comparing the gene expression of selected groups of cells from one or more preprocessed Seurat objects
}
